% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{action_filter_data}
\alias{action_filter_data}
\title{Filter Data}
\usage{
action_filter_data(g, input, asp, operator = ">")
}
\arguments{
\item{g}{An object of class \code{g2r} or \code{g2Proxy} as returned by \code{\link[=g2]{g2()}} or
\code{\link[=g2_proxy]{g2_proxy()}}.}

\item{input}{The \code{id} of the input that triggers the filter,
either \code{\link[=input_select]{input_select()}} or \code{\link[=input_slider]{input_slider()}}.}

\item{asp}{Aspect (column) to filter.}

\item{operator}{Operator of the filter, this is combined with
the value from the \code{input} and the \code{asp} to form a filter
statement with the following template; \verb{asp operator inputValue}.
For instance, a filter on input \code{id = "theFilter"} on the column
\code{speed} (of the \code{cars} dataset) with the operator \code{>} (greater than)
will create the following filter statement: \code{speed > inputValue}}
}
\description{
Filter data.
}
\examples{
# works in Rmarkdown
input_slider(
  "yFilter",
  "Filter Y >",
  value = 50,
  min = 40,
  max = 70,
  step = 5
)

\dontrun{
g2(cars, asp(speed, dist)) \%>\%
  fig_point() \%>\%
  fig_smooth() \%>\%
  action_filter_data(
    "yFilter",
    dist,
    operator = ">"
  )
}
}
