% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{action_toggle_visibility}
\alias{action_toggle_visibility}
\title{Toggle Visibility}
\usage{
action_toggle_visibility(g, btn)
}
\arguments{
\item{g}{An object of class \code{g2r} or \code{g2Proxy} as returned by \code{\link[=g2]{g2()}} or
\code{\link[=g2_proxy]{g2_proxy()}}.}

\item{btn}{Id of the \link{input_button} that toggles
the visibility.}
}
\description{
Toggle the visibily of a chart.
}
\examples{
# works in Rmarkdown
input_button("toggle", "Show/hide chart")

g2(mtcars, asp(qsec, mpg)) \%>\% 
 fig_point() \%>\% 
 action_toggle_visibility("toggle")

}
