% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.R
\name{adjust}
\alias{adjust}
\title{Adjust}
\usage{
adjust(type, margin = NULL, dodge_by = NULL)
}
\arguments{
\item{type}{A vector of types of adjustement to apply
to the figure, see the "types" section below for valid values.}

\item{margin}{Margin, between \code{0} and \code{1}.}

\item{dodge_by}{Bare column name to use as group for dodge.}
}
\description{
Adjust a figure.
}
\section{Types}{

Valid values for the \code{type} argument.
\itemize{
\item \code{stack}
\item \code{dodge}
\item \code{jitter}
\item \code{symmetric}
}
}

\examples{
df <- data.frame(
  x = c(letters, letters),
  y = runif(52),
  grp = rep(c("A", "Z"), each = 2)
)

g2(df, asp(x, y, color = grp)) \%>\%
  fig_interval(adjust("stack"))
}
