% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R
\name{aka}
\alias{aka}
\title{Aliases}
\usage{
aka(g, asp, alias)
}
\arguments{
\item{g}{An object of class \code{g2r} or \code{g2Proxy} as returned by \code{\link[=g2]{g2()}} or
\code{\link[=g2_proxy]{g2_proxy()}}.}

\item{asp}{Bare name of aspect to alias.}

\item{alias}{A string defining the alias.}
}
\description{
Aliases an aspect, this changes the name of
the aspects when displayed in axis titles,
tooltips, labels, and other places.
}
\examples{
# see tooltip
g2(cars, asp(speed, dist)) \%>\% 
 fig_point() \%>\% 
 aka(speed, "SO FAST")

}
