% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{alter}
\alias{alter}
\title{Dataset}
\usage{
alter(data, ..., source = NULL, .rename = TRUE, .rows = TRUE,
  .return = NULL)
}
\arguments{
\item{data}{A data.frame to alter.}

\item{...}{Altering options.}

\item{source}{A \code{list} of source options.}

\item{.rename}{Whether to rename the columns of the altered data.frame (recommended).}

\item{.rows}{Wheter to return the whole altered object or just the rows.}

\item{.return}{Method to return, i.e.\code{"rows"}.}
}
\description{
Dataset with state management \url{https://github.com/antvis/data-set}
}
\examples{
head(fruits)

fr <- fruits \%>\% 
  dplyr::mutate(value = value * 100) \%>\% 
  dplyr::select(name = fruit, value) \%>\% 
  alter(type = "waffle", maxCount = 500, rows = 12) 

head(fr)

g2(fr, asp(x, y, shape = "square", size = 10, color = name)) \%>\% 
  fig_point(axes = FALSE) \%>\% 
  gauge_axis(x, FALSE)
}
