% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animation.R
\name{animation}
\alias{animation}
\alias{Animation}
\title{Animations}
\description{
Helper function to build animations.
}
\examples{
# create animation
anim <- Animation$
  new()$
  appear(
  duration = 2000,
  delay = 500
)

iris \%>\%
  g2(asp(Sepal.Length, Sepal.Width, color = Species)) \%>\%
  fig_point(anim)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-enter}{\code{Animation$enter()}}
\item \href{#method-leave}{\code{Animation$leave()}}
\item \href{#method-appear}{\code{Animation$appear()}}
\item \href{#method-update}{\code{Animation$update()}}
\item \href{#method-print}{\code{Animation$print()}}
\item \href{#method-retrieve}{\code{Animation$retrieve()}}
\item \href{#method-clone}{\code{Animation$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-enter"></a>}}
\if{latex}{\out{\hypertarget{method-enter}{}}}
\subsection{Method \code{enter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Animation$enter(animation = NULL, easing = NULL, delay = NULL, duration = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{animation}}{Name of animation; \code{clipIn}, \code{zoomIn},
\code{pathIn}, \code{scaleInY}, \code{scaleInX}, \code{fanIn}, or \code{fadeIn}.}

\item{\code{easing}}{Name of easing function.}

\item{\code{delay, duration}}{Delay and duration in milliseconds.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Animation to use on enter
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-leave"></a>}}
\if{latex}{\out{\hypertarget{method-leave}{}}}
\subsection{Method \code{leave()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Animation$leave(animation = NULL, easing = NULL, delay = NULL, duration = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{animation}}{Name of animation; \code{lineWidthOut},
\code{zoomOut}, \code{pathOut}, or \code{fadeOut}.}

\item{\code{easing}}{Name of easing function.}

\item{\code{delay, duration}}{Delay and duration in milliseconds.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Animation to use on leave
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-appear"></a>}}
\if{latex}{\out{\hypertarget{method-appear}{}}}
\subsection{Method \code{appear()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Animation$appear(
  animation = NULL,
  easing = NULL,
  delay = NULL,
  duration = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{animation}}{Name of animation; \code{clipIn},
\code{zoomIn}, \code{pathIn}, \code{scaleInY}, \code{scaleInX}, \code{fanIn},
or \code{fadeIn}.}

\item{\code{easing}}{Name of easing function.}

\item{\code{delay, duration}}{Delay and duration in milliseconds.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Animation to use on appear
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update"></a>}}
\if{latex}{\out{\hypertarget{method-update}{}}}
\subsection{Method \code{update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Animation$update(
  animation = NULL,
  easing = NULL,
  delay = NULL,
  duration = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{animation}}{Name of animation; \code{fadeIn}, or \code{fanIn}.}

\item{\code{easing}}{Name of easing function.}

\item{\code{delay, duration}}{Delay and duration in milliseconds.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Animation to use on appear
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Animation$print()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Print
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-retrieve"></a>}}
\if{latex}{\out{\hypertarget{method-retrieve}{}}}
\subsection{Method \code{retrieve()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Animation$retrieve()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Retrieve the animation list
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Animation$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
