% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_widget}
\alias{as_widget}
\title{Create widget}
\usage{
as_widget(x, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{x}{List of options.}

\item{width, height}{Dimensions of chart.}

\item{elementId}{Id of element to create.}
}
\description{
Create widget.
}
\keyword{internal}
