% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspects.R
\name{asp}
\alias{asp}
\title{Aspects}
\usage{
asp(x, y, ...)
}
\arguments{
\item{x, y}{Defines values to map to cartesian coordinates.}

\item{...}{Any other key value pair of aspect.}
}
\description{
Define aspects of a visualisation.
}
\section{Aspects}{

\itemize{
\item \code{x}, \code{y}
\item \code{size}
\item \code{color}
\item \code{shape}
\item \code{label}
}
}

