% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aspect_action}
\alias{aspect_action}
\title{Pass an action bound to an aspect}
\usage{
aspect_action(g, asps, ..., action)
}
\description{
Pass an action bound to an aspect
}
\section{Uses}{

\itemize{
\item \code{legend_asps}
\item \code{axis_asps}
}
}

\keyword{internal}
