% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{title}
\alias{title}
\alias{eXAxis}
\alias{eYAxis}
\title{Create an axis for a chart}
\usage{
title(chart, visible = TRUE, text = NULL, style = NULL)

eXAxis(chart, ...)

eYAxis(chart, ...)
}
\description{
Add an axis to a chart.
}
\details{
This function modified a few default options for the axis component in
ECharts: 1) \code{scale = TRUE} (was \code{FALSE} by default in ECharts); 2)
\code{axisLine$onZero = FALSE} (was \code{TRUE} in ECharts).
}
