% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{axis_label}
\alias{axis_label}
\alias{axis_label_x}
\alias{axis_label_x.g2r}
\alias{axis_label_y}
\alias{axis_label_y.g2r}
\title{Axis Label}
\usage{
axis_label_x(g, label, ..., offset = 30)

\method{axis_label_x}{g2r}(g, label, ..., offset = 30)

axis_label_y(g, label, ..., offset = 30)

\method{axis_label_y}{g2r}(g, label, ..., offset = 30)
}
\arguments{
\item{g}{An object of class \code{g2r} or \code{g2Proxy} as returned by \code{\link[=g2]{g2()}} or
\code{\link[=g2_proxy]{g2_proxy()}}.}

\item{label}{Label to use on the axis.}

\item{...}{Options to customise the title.}

\item{offset}{Offset between label and axis,
if \code{0} the label is not visible.}
}
\description{
Add axis labels.
}
