% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{axis_label_}
\alias{axis_label_}
\title{Axis Label}
\usage{
axis_label_(g, label, ..., offset = 30, axis = c("x", "y"))
}
\arguments{
\item{g}{An object of class \code{g2r} or \code{g2Proxy} as returned by \code{\link[=g2]{g2()}} or
\code{\link[=g2_proxy]{g2_proxy()}}.}

\item{label}{Label to use.}

\item{...}{Options passed to the \code{title} argument}

\item{offset}{Label offset, otherwise (if \code{0}),
not visible.}

\item{axis}{Axis to put the label on.}
}
\description{
Axis Label
}
\keyword{internal}
