% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{axis_title}
\alias{axis_title}
\alias{axis_title_x}
\alias{axis_title_x.g2r}
\alias{axis_title_y}
\alias{axis_title_y.g2r}
\title{Axis Title}
\usage{
axis_title_x(g, title, ..., fontSize = 10, offset = 30)

\method{axis_title_x}{g2r}(g, title, ..., fontSize = 10, offset = 30)

axis_title_y(g, title, ..., fontSize = 10, offset = 30)

\method{axis_title_y}{g2r}(g, title, ..., fontSize = 10, offset = 30)
}
\arguments{
\item{g}{An object of class \code{g2r} or \code{g2Proxy} as returned by \code{\link[=g2]{g2()}} or
\code{\link[=g2_proxy]{g2_proxy()}}.}

\item{title}{Title to use on the axis.}

\item{...}{Options to customise the title.}

\item{offset}{Offset between title and axis,
if \code{0} the title is not visible.}
}
\description{
Add axis titles.
}
