% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{axis_title_}
\alias{axis_title_}
\title{Axis Label}
\usage{
axis_title_(g, title, ..., fontSize = 10, offset = 30, axis = c("x", "y"))
}
\arguments{
\item{g}{An object of class \code{g2r} or \code{g2Proxy} as returned by \code{\link[=g2]{g2()}} or
\code{\link[=g2_proxy]{g2_proxy()}}.}

\item{title}{Title to use.}

\item{...}{Options passed to the \code{title} argument}

\item{offset}{Title offset, otherwise (if \code{0}),
not visible.}

\item{axis}{Axis to put the title on.}
}
\description{
Axis Label
}
\keyword{internal}
