% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{capture_event}
\alias{capture_event}
\title{Events}
\usage{
capture_event(g, event, callback = NULL, when = c("on", "once", "off"))
}
\arguments{
\item{g}{An object of class \code{g2r} or \code{g2Proxy} as returned by \code{\link[=g2]{g2()}} or
\code{\link[=g2_proxy]{g2_proxy()}}.}

\item{event}{Name of event to trigger the callback.}

\item{callback}{A callback function to run when the \code{event}
is fired, if \code{NULL} then a default one is created, see
details.}

\item{when}{When the event should be triggered.}
}
\description{
Capture events in shiny.
}
\details{
The \code{callback} function should accept a single
argument; the event data. If no callback function is passed
(\code{NULL}) then one is generated. The generated callback function
sets a shiny input \verb{<chartId>_<eventName>} with the event data.
}
\examples{
g2(iris, asp(Sepal.Width, Sepal.Length)) \%>\%
 fig_point() \%>\% 
 capture_event("point:click")

}
