% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxies.R
\name{change-size}
\alias{change-size}
\alias{change_size}
\alias{change_width}
\alias{change_height}
\title{Change Size}
\usage{
change_size(proxy, width, height)

change_width(proxy, width)

change_height(proxy, height)
}
\arguments{
\item{proxy}{A g2r proxy as returned by \code{\link{g2Proxy}}}

\item{width, height}{New dimensions of chart.}
}
\description{
Change size of chart.
}
\examples{
library(shiny)

ui <- fluidPage(
  sliderInput("width", "width", min = 100, max = 700, value = 250),
  g2Output("chart")
)

server <- function(input, output) {
  output$chart <- renderG2({
    g2(cars, asp(speed, dist)) \%>\% 
      fig_point()
  })

  observeEvent(input$width, {
    g2Proxy("chart") \%>\% 
      change_width(input$width)
  })
}

if(interactive()) shinyApp(ui, server)

}
