% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{change_data}
\alias{change_data}
\title{Change the data}
\usage{
change_data(g, data)
}
\arguments{
\item{g}{An object of class \code{g2r} or \code{g2Proxy} as returned by \code{\link[=g2]{g2()}} or
\code{\link[=g2_proxy]{g2_proxy()}}.}

\item{data}{New dataset to replaced the one used
to currently plot the data.}
}
\description{
Dynamically change the data of a shiny plot.
}
\examples{
library(shiny)

makeData <- function(){
 data.frame(
   x = runif(100),
   y = runif(100),
   size = runif(100)
 )
}

ui <- fluidPage(
 g2Output("plot"),
 actionButton("change", "Change data")
)

server <- function(input, output){

 output$plot <- renderG2({
   g2(makeData(), asp(x, y, size = size)) \%>\% 
     fig_point()
 })

 observeEvent(input$change, {
   g2_proxy("plot") \%>\% 
     change_data(makeData())
 })
}

if(interactive()){
 shinyApp(ui, server)
}

}
