% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxies.R
\name{chart-proxies}
\alias{chart-proxies}
\alias{repaint}
\alias{clear}
\alias{render}
\alias{destroy}
\title{Manipulate Chart}
\usage{
repaint(proxy)

clear(proxy)

render(proxy)

destroy(proxy)
}
\arguments{
\item{proxy}{A g2r proxy as returned by \code{\link{g2Proxy}}}
}
\description{
Repaint, clear, render, or destroy the chart.
}
