% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{chart-title}
\alias{chart-title}
\alias{chart_title}
\alias{chart_subtitle}
\title{Title & Subtitle}
\usage{
chart_title(g2, ..., position = c("top", "bottom"),
  tag = htmltools::h1, inherit_padding = TRUE, font_family = NULL)

chart_subtitle(g2, ..., position = c("top", "bottom"),
  tag = htmltools::h4, inherit_padding = TRUE, font_family = NULL)
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}

\item{...}{Title or subtitle content and style, passed to \code{tag}.}

\item{position}{Position of title or subtitle, \code{top} or \code{bottom}.}

\item{tag}{A valid html tag \link[htmltools]{tag}.}

\item{inherit_padding}{Whether to inherits the \code{left} padding set on chart object (\code{g2}).}

\item{font_family}{Font family to use, if \code{NULL} sets to g2 default.}
}
\description{
Add a chart title and subtitle.
}
