% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspects.R
\name{collapse_asp}
\alias{collapse_asp}
\title{Collapse Aspects}
\usage{
collapse_asp(asp)
}
\arguments{
\item{asp}{Aspects, a character vector to collapse}
}
\description{
Collapse aspects on a single line. Only \code{position}
works with a vector/array (\verb{["x", "y"]}), hence enforcement
of \code{x*y} everywhere.
}
\examples{
\dontrun{
collapse_asp(c("x", "y"))
}

}
\keyword{internal}
