% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{conf_legend}
\alias{conf_legend}
\title{Legend}
\usage{
conf_legend(g2, ...)
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}

\item{...}{Configuration options or a logical indicating whether to show the legend.}
}
\description{
Configure the chart's legend.
}
\examples{
g <- g2(cars, plan(speed, dist, color = dist)) \%>\% 
  fig_point()

g \%>\% conf_legend(FALSE)

g \%>\% conf_legend(position = "right")

}
