% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{conf_renderer}
\alias{conf_renderer}
\title{Renderer}
\usage{
conf_renderer(g2, renderer = c("svg", "canvas"))
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}

\item{renderer}{Renderer, \code{canvas} or \code{svg}.}
}
\description{
Define renderer.
}
\note{
The "g" in g2r stands for svg.
}
