% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltip.R
\name{conf_tooltip}
\alias{conf_tooltip}
\title{Configure Tooltip}
\usage{
conf_tooltip(g, ...)
}
\arguments{
\item{g}{An object of class \code{g2r} or \code{g2Proxy} as returned by \code{\link[=g2]{g2()}} or
\code{\link[=g2_proxy]{g2_proxy()}}.}

\item{...}{Options to pass to the axis, pass \code{FALSE}
to hide the axis. Visit the
\href{https://g2.antv.vision/en/docs/api/general/tooltip}{official documentation}
for the full list of options.}
}
\description{
Configure the tooltip applied to the entire chart.
See \code{\link[=gauge_tooltip]{gauge_tooltip()}} to customise a specific tooltip
(the tooltip of a specific figure).
}
\examples{
g2(mtcars, asp(drat, qsec, color = hp)) \%>\% 
 fig_point() \%>\% 
 conf_tooltip(
   showCrosshairs = TRUE,
   crosshairs = list(type = "xy")
)

}
