% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config}
\alias{config}
\title{Configure Figure}
\usage{
config(
  id = NULL,
  visible = NULL,
  region = NULL,
  padding = NULL,
  theme = NULL,
  ...
)
}
\arguments{
\item{id}{Id of figure.}

\item{visible}{Whether the figure is visible.}

\item{region}{Region that the figure should occupy
on the canvas.}

\item{padding}{Padding around the figure.}

\item{theme}{Theme of the figure.}

\item{...}{Any other options from the
\href{https://g2.antv.vision/en/docs/api/general/chart#viewcfg}{official documentation}.}
}
\description{
Configure a figure.
}
\examples{
mtcars \%>\%
  g2(asp(qsec)) \%>\%
  fig_point(
    asp(y = mpg),
    config(
      region = list(
        start = list(x = 0, y = 0),
        end = list(x = 0.5, y = 1)
      )
    )
  ) \%>\%
  fig_point(
    asp(y = wt),
    config(
      region = list(
        start = list(x = 0.5, y = 0),
        end = list(x = 1, y = 1)
      )
    )
  )
}
