% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxies.R
\name{convert}
\alias{convert}
\alias{to_dataURL}
\alias{download_image}
\title{Convert}
\usage{
to_dataURL(proxy)

download_image(proxy, name = "g2r")
}
\arguments{
\item{proxy}{A g2r proxy as returned by \code{\link{g2Proxy}}}

\item{name}{Name of file.}
}
\description{
Convert chart to a dataURL or an image.
}
\examples{
library(shiny)

ui <- fluidPage(
  actionButton("dl", "download"),
  g2Output("chart")
)

server <- function(input, output) {
  output$chart <- renderG2({
    g2(cars, asp(speed, dist)) \%>\% 
      fig_point()
  })

  observeEvent(input$dl, {
    g2Proxy("chart") \%>\% 
      download_image()
  })
}
\dontrun{shinyApp(ui, server)}

}
