% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstalk.R
\name{crosstalk_select}
\alias{crosstalk_select}
\title{Crosstalk Customisation}
\usage{
crosstalk_select(g, attribute, on, off = NULL)
}
\arguments{
\item{g}{An object of class \code{g2r} or \code{g2Proxy} as returned by \code{\link[=g2]{g2()}} or
\code{\link[=g2_proxy]{g2_proxy()}}.}

\item{attribute}{Attribute to customise, e.g.: \code{stroke}, \code{fill}, \code{strokeOpacity}, etc.}

\item{on, off}{Value to set the \code{attribute} to if the selected is on or off.}
}
\description{
Customise the crosstalk selection handle.
}
