% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{default_callback}
\alias{default_callback}
\title{Default Callback}
\usage{
default_callback(event, callback = NULL)
}
\arguments{
\item{event}{Name of event.}

\item{callback}{Callback function, or \code{NULL}.}
}
\description{
Return a default callback in the event none were
specified.
}
\keyword{internal}
