% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{dependencies}
\alias{dependencies}
\alias{dep_g2}
\alias{dep_dataset}
\alias{dep_bizcharts}
\alias{dep_corejs}
\title{Dependencies for g2, bizcharts, and data-set}
\usage{
dep_g2()

dep_dataset()

dep_bizcharts()

dep_corejs()
}
\value{
\code{htmltools::htmlDependency}
}
\description{
Dependencies for g2, bizcharts, and data-set
}
