% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{drop_nulls}
\alias{drop_nulls}
\title{Remove NULLs}
\usage{
drop_nulls(x)
}
\arguments{
\item{x}{A \code{list}.}
}
\description{
Remove NULLs
}
\examples{
\dontrun{
drop_nulls(list(NULL, 1))
}

}
\keyword{internal}
