% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{elements}
\alias{elements}
\alias{element}
\alias{elementPoint}
\alias{elementLine}
\alias{elementArea}
\alias{elementEdge}
\alias{elementInterval}
\alias{elementPolygon}
\alias{elementSchema}
\title{Element}
\usage{
element(
  ...,
  shape = NULL,
  figure = c("point", "area", "edge", "line", "interval", "polygon", "schema"),
  state = c("default", "active", "inactive", "selected")
)

elementPoint(
  ...,
  shape = c("hollow-circle", "cross", "hypen", "line", "plus", "tick", "circle",
    "square", "bowtie", "diamond", "hexagon", "triangle", "triangle-down",
    "hollow-square", "hollow-bowtie", "hollow-triangle-down"),
  state = c("default", "active", "inactive", "selected")
)

elementLine(
  ...,
  shape = c("line", "dot", "dash", "smooth", "hv", "vh", "hvh", "vhv"),
  state = c("default", "active", "inactive", "selected")
)

elementArea(
  ...,
  shape = c("area", "smooth", "line", "smooth-line"),
  state = c("default", "active", "inactive", "selected")
)

elementEdge(
  ...,
  shape = c("line", "vhv", "smooth", "arc"),
  state = c("default", "active", "inactive", "selected")
)

elementInterval(
  ...,
  shape = c("rect", "hollow-rect", "line", "tick", "funnel", "pyramid"),
  state = c("default", "active", "inactive", "selected")
)

elementPolygon(..., state = c("default", "active", "inactive", "selected"))

elementSchema(
  ...,
  shape = c("box", "candle"),
  state = c("default", "active", "inactive", "selected")
)
}
\arguments{
\item{...}{Key value pairs to pass to \code{style}.}

\item{shape}{Shape to modify, if \code{NULL} selects
a common default based on the figure,
e.g.: \code{hollow-circle} for the \code{point} shape.}

\item{figure}{Figure to modify.}

\item{state}{State of the shape to modify.}
}
\description{
Function to use in \code{\link[=motif]{motif()}} and style specific elements.
}
\section{Functions}{


\code{\link[=element]{element()}} will work for any figure, but other
functions may be more convienient to use.
\itemize{
\item \code{element}: Customise any element.
\item \code{elementPoint}: Customise point.
\item \code{elementLine}: Customise line.
\item \code{elementArea}: Customise area.
\item \code{elementEdge}: Customise edge.
\item \code{elementInterval}: Customise interval.
\item \code{elementPolygon}: Customise polygon.
\item \code{elementSchema}: Customise schema.
}
}

\examples{
g2(iris, asp(Sepal.Width, Sepal.Length)) \%>\%
  fig_point(
    asp(color = Species, shape = "circle")
  ) \%>\%
  motif(
    brandColor = "orange",
    backgroundColor = "black",
    elementPoint(
      shape = "circle",
      stroke = "white",
      fillOpacity = .7
    )
  )
}
