% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{fig_bin}
\alias{fig_bin}
\title{Bin}
\usage{
fig_bin(
  g,
  ...,
  type = c("rectangle", "hexagon"),
  bins = c(10, 10),
  size_count = TRUE,
  sync = TRUE,
  data = NULL,
  inherit_asp = TRUE
)
}
\arguments{
\item{g}{An object of class \code{g2r} or \code{g2Proxy} as returned by \code{\link[=g2]{g2()}} or
\code{\link[=g2_proxy]{g2_proxy()}}.}

\item{...}{Options to pass to the figure, including \code{\link[=asp]{asp()}},
and \code{\link[=adjust]{adjust()}}, \code{\link[=active]{active()}}, \code{\link[=selected]{selected()}}, and \code{\link[=config]{config()}},
other key value pairs of are passed to
\href{https://g2.antv.vision/en/docs/api/general/style}{style}.}

\item{type}{The shape of bin to create.}

\item{bins}{Number of bins by dimension (width, height).}

\item{size_count}{Whether to size the binds by count.}

\item{sync}{Whether to sync the axis data (align) with that
used in other figures, set to \code{FALSE} to not sync or set to
a character string to use as name of sync group.}

\item{data}{A dataset (\code{data.frame} or \code{tibble}) to use to
draw the figure.}

\item{inherit_asp}{Whether to inherit the aspects paseed to
\code{\link[=g2]{g2()}} initialisation function.}
}
\description{
Add a bin figure to the chart.
}
\details{
Requires the \code{x} and \code{y} aspects.
}
\examples{
g2(cars, asp(speed, dist)) \%>\% 
 fig_bin(size_count = FALSE)

g2(cars, asp(speed, dist)) \%>\% 
 fig_bin(type = "hexagon")

}
