% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{fig_candle}
\alias{fig_candle}
\title{Candle}
\usage{
fig_candle(g, ..., sync = TRUE, data = NULL, inherit_asp = TRUE)
}
\arguments{
\item{g}{An object of class \code{g2r} or \code{g2Proxy} as returned by \code{\link[=g2]{g2()}} or
\code{\link[=g2_proxy]{g2_proxy()}}.}

\item{...}{Options to pass to the figure, including \code{\link[=asp]{asp()}},
and \code{\link[=adjust]{adjust()}}, \code{\link[=active]{active()}}, \code{\link[=selected]{selected()}}, and \code{\link[=config]{config()}},
other key value pairs of are passed to
\href{https://g2.antv.vision/en/docs/api/general/style}{style}.}

\item{sync}{Whether to sync the axis data (align) with that
used in other figures, set to \code{FALSE} to not sync or set to
a character string to use as name of sync group.}

\item{data}{A dataset (\code{data.frame} or \code{tibble}) to use to
draw the figure.}

\item{inherit_asp}{Whether to inherit the aspects paseed to
\code{\link[=g2]{g2()}} initialisation function.}
}
\description{
Add a candle figure to the chart.
}
\details{
Requires the following aspects defined:
\itemize{
\item \code{open}
\item \code{close}
\item \code{high}
\item \code{low}
}

If no \code{color} argument is passed the candles are colored
according to their trend (open > close = "up").
}
\examples{
stock <- structure(list(date = structure(c(18626, 18627, 18631, 18632), class = "Date"), 
 open = c(39.52, 39.330002, 40.169998, 41.5), high = c(39.73, 
 40, 41.560001, 42.040001), low = c(39.200001, 39.029999, 
 39.939999, 40.77), close = c(39.34, 39.880001, 41.400002, 
 41.16)), row.names = c(NA, -4L), class = c("tbl_df", "tbl", 
 "data.frame")
)

g2(stock, asp(date, open = open, close = close, high = high, low = low)) \%>\% 
 fig_candle() \%>\% 
 gauge_x_time_cat()

}
