% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{fig_contour}
\alias{fig_contour}
\title{Contour}
\usage{
fig_contour(
  g,
  ...,
  sync = TRUE,
  data = NULL,
  inherit_asp = TRUE,
  colors = NULL,
  nlevels = 10,
  binwidth,
  levels,
  criticalRatio = 5
)
}
\arguments{
\item{g}{An object of class \code{g2r} or \code{g2Proxy} as returned by \code{\link[=g2]{g2()}} or
\code{\link[=g2_proxy]{g2_proxy()}}.}

\item{...}{Options to pass to the figure, including \code{\link[=asp]{asp()}},
and \code{\link[=adjust]{adjust()}}, \code{\link[=active]{active()}}, \code{\link[=selected]{selected()}}, and \code{\link[=config]{config()}},
other key value pairs of are passed to
\href{https://g2.antv.vision/en/docs/api/general/style}{style}.}

\item{sync}{Whether to sync the axis data (align) with that
used in other figures, set to \code{FALSE} to not sync or set to
a character string to use as name of sync group.}

\item{data}{A dataset (\code{data.frame} or \code{tibble}) to use to
draw the figure.}

\item{inherit_asp}{Whether to inherit the aspects paseed to
\code{\link[=g2]{g2()}} initialisation function.}

\item{colors}{A palette of colors to define the \code{stroke}
of each path.}

\item{nlevels}{Passed to \code{contoureR::getContourLines}.
An integer number of bins to split the data into \emph{iff}
‘levels’ or ‘binwidth’ have not been specified.}

\item{binwidth}{Passed to \code{contoureR::getContourLines}.
The desired width of the bins, if specified, will override
‘nlevels’}

\item{levels}{Passed to \code{contoureR::getContourLines}.
A numeric vector of the explicitly specified levels (zvalues)
to contour, by specifying this argument, it will override
‘nlevels’ and/or ‘binwidth’. If this argument is provided,
the stacking order of the contours will be preserved in the
order of first occurence within the supplied vector.}

\item{criticalRatio}{Passed to \code{contoureR::getContourLines}.
When producing the Delaunay Mesh, the quality of the mesh can
be poor in the proximity to the convex hull, Del's that have
an aspect ratio greater than this value are not considered when
producing the contours. In this context, the aspect ratio is
defined as the circumradius to twice its inradius, equilateral
triangles have an aspect ratio of 1, everything else is larger}
}
\description{
Add a contour line figure to the chart.
}
\details{
Requires the \code{x}, \code{y} and \code{z} aspects, the
width of the error bars can be changed with the \code{size}
aspect.
}
\examples{
data(volcano)

x <- 1:nrow(volcano)
y <- 1:ncol(volcano)
df <- expand.grid(x = x,y = y)
df$z = apply(df, 1, function(x){ 
 volcano[x[1],x[2]]
})

g <- g2(df, asp(x, y, z = z)) 

fig_contour(g)

fig_contour(g, colors = c("red", "blue"))

}
