% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{fig_dotplot}
\alias{fig_dotplot}
\title{Dot plot}
\usage{
fig_dotplot(g2, ..., count = 500, dot_size = list(1, 1),
  gap_ratio = 0.1, data = NULL, inherit_asp = TRUE, name = NULL)
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}

\item{...}{Any option, aesthetic (\code{\link{asp}}), or animation (\code{\link{Animation}}).}

\item{count}{Number of dots to use.}

\item{dot_size}{Size of dots.}

\item{gap_ratio}{Gap between dots.}

\item{data}{A \code{data.frame} containing data to chart or a \code{list}.}

\item{inherit_asp}{Whether to inherit aspects from \code{g2r}.}

\item{name}{Name of figure, useful to apply functions to specific figures.}
}
\description{
Create a dot plot based on \code{x}, \code{y}, and optionally \code{color} aspects.
}
\examples{
g2(states, asp(name, value, color = State)) \%>\% 
  fig_dotplot(asp(shape = "square")) 

}
