% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{fig_error_bar}
\alias{fig_error_bar}
\title{Error bar}
\usage{
fig_error_bar(g2, ..., line_size = 1, tip_size = 5, data = NULL,
  inherit_asp = TRUE, name = NULL)
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}

\item{...}{Any option, aesthetic (\code{\link{asp}}), or animation (\code{\link{Animation}}).}

\item{line_size, tip_size}{Defines size of error bars.}

\item{data}{A \code{data.frame} containing data to chart or a \code{list}.}

\item{inherit_asp}{Whether to inherit aspects from \code{g2r}.}

\item{name}{Name of figure, useful to apply functions to specific figures.}
}
\description{
Add error bars based \code{y}, \code{ymin}, \code{ymax}, and optionally \code{group} aspects.
}
\examples{
df <- data.frame(
  trt = factor(c(1, 1, 2, 2)),
  resp = c(1, 5, 3, 4),
  group = factor(c(1, 2, 1, 2)),
  upper = c(1.1, 5.3, 3.3, 4.2),
  lower = c(0.8, 4.6, 2.4, 3.6)
)

g2(df, asp(trt, ymin = lower, ymax = upper, color = group)) \%>\% 
  fig_error_bar()

}
