% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{fig_histogram}
\alias{fig_histogram}
\title{Histogram}
\usage{
fig_histogram(
  g,
  ...,
  bin_width = 5,
  sync = TRUE,
  data = NULL,
  inherit_asp = TRUE
)
}
\arguments{
\item{g}{An object of class \code{g2r} or \code{g2Proxy} as returned by \code{\link[=g2]{g2()}} or
\code{\link[=g2_proxy]{g2_proxy()}}.}

\item{...}{Options to pass to the figure, including \code{\link[=asp]{asp()}},
and \code{\link[=adjust]{adjust()}}, \code{\link[=active]{active()}}, and \code{\link[=selected]{selected()}}, other key value
pairs of are passed to
\href{https://g2.antv.vision/en/docs/api/general/style}{style}.}

\item{bin_width}{Width of bin.}

\item{sync}{Whether to sync the axis data (align) with that
used in other figures, set to \code{FALSE} to not sync or set to
a character string to use as name of sync group.}

\item{data}{A dataset (\code{data.frame} or \code{tibble}) to use to
draw the figure.}

\item{inherit_asp}{Whether to inherit the aspects paseed to
\code{\link[=g2]{g2()}} initialisation function.}
}
\description{
Add a histogram figure to the chart.
}
\examples{
df <- data.frame(
 grp = rep(c("A", "B"), each = 200),
   val = c(
     rnorm(200, mean = 57, sd = 5), 
     rnorm(200, mean = 53, sd = 5)
   )
)

g2(df, asp(val, color = grp)) \%>\% 
 fig_histogram(adjust("stack"), bin_width = 1)

}
