% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{fig_line}
\alias{fig_line}
\title{Line}
\usage{
fig_line(g, ..., sync = TRUE, data = NULL, inherit_asp = TRUE, style = NULL)
}
\arguments{
\item{g}{An object of class \code{g2r} as returned by \code{\link[=g2]{g2()}}.}

\item{...}{Options to pass to the figure, including \code{\link[=asp]{asp()}},
and \code{\link[=adjust]{adjust()}}.}

\item{sync}{Whether to sync the axis data (align).}

\item{data}{A dataset (\code{data.frame} or \code{tibble}) to use to
draw the figure.}

\item{inherit_asp}{Whether to inherit the aspects paseed to
\code{\link[=g2]{g2()}} initialisation function.}

\item{style}{A list of options to pass to
\href{https://g2.antv.vision/en/docs/api/general/style}{style} that
customise the figure.}
}
\description{
Add a line figure.
}
\examples{
g2(CO2, asp(conc, uptake, color = Plant)) \%>\% 
 fig_line()

}
