% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{fig_map}
\alias{fig_map}
\title{Map}
\usage{
fig_map(g, ..., map = "World")
}
\arguments{
\item{g}{An object of class \code{g2r} or \code{g2Proxy} as returned by \code{\link[=g2]{g2()}} or
\code{\link[=g2_proxy]{g2_proxy()}}.}

\item{...}{Options to pass to the figure, including \code{\link[=asp]{asp()}},
and \code{\link[=adjust]{adjust()}}, \code{\link[=active]{active()}}, \code{\link[=selected]{selected()}}, and \code{\link[=config]{config()}},
other key value pairs of are passed to
\href{https://g2.antv.vision/en/docs/api/general/style}{style}.}

\item{map}{Name of map, or map object as returned by
\code{\link[=get_map_data]{get_map_data()}}, or a \code{SpatialPolygonsDataFrame} as
as returned by \link[raster:getData]{raster::getData}, or a \code{geo_list} as
obtained from \code{geojsonio::geojson_list()}.}
}
\description{
Add a map figure.
}
