% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{fig_point}
\alias{fig_point}
\title{Point}
\usage{
fig_point(g, ..., sync = TRUE, data = NULL, inherit_asp = TRUE)
}
\arguments{
\item{g}{An object of class \code{g2r} or \code{g2Proxy} as returned by \code{\link[=g2]{g2()}} or
\code{\link[=g2_proxy]{g2_proxy()}}.}

\item{...}{Options to pass to the figure, including \code{\link[=asp]{asp()}},
and \code{\link[=adjust]{adjust()}}, \code{\link[=active]{active()}}, \code{\link[=selected]{selected()}}, and \code{\link[=config]{config()}},
other key value pairs of are passed to
\href{https://g2.antv.vision/en/docs/api/general/style}{style}.}

\item{sync}{Whether to sync the axis data (align) with that
used in other figures, set to \code{FALSE} to not sync or set to
a character string to use as name of sync group.}

\item{data}{A dataset (\code{data.frame} or \code{tibble}) to use to
draw the figure.}

\item{inherit_asp}{Whether to inherit the aspects paseed to
\code{\link[=g2]{g2()}} initialisation function.}
}
\description{
Add a point figure.
}
\examples{
g2(cars) \%>\%
  fig_point(asp(speed, dist))

g2(mtcars, asp(mpg, disp, size = qsec)) \%>\%
  fig_point(asp(color = "red", shape = "square"))
}
