% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{fig_polygon}
\alias{fig_polygon}
\title{Polygon}
\usage{
fig_polygon(g, ..., sync = TRUE, data = NULL, inherit_asp = TRUE)
}
\arguments{
\item{g}{An object of class \code{g2r} as returned by \code{\link[=g2]{g2()}}.}

\item{...}{Options to pass to the figure, including \code{\link[=asp]{asp()}},
and \code{\link[=adjust]{adjust()}}.}

\item{sync}{Whether to sync the axis data (align).}

\item{data}{A dataset (\code{data.frame} or \code{tibble}) to use to
draw the figure.}

\item{inherit_asp}{Whether to inherit the aspects paseed to
\code{\link[=g2]{g2()}} initialisation function.}
}
\description{
Add a polygon figure.
}
