% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{fig_primitive}
\alias{fig_primitive}
\title{Primitive}
\usage{
fig_primitive(
  g,
  ...,
  data = NULL,
  inherit_asp = TRUE,
  sync = TRUE,
  type = c("point", "interval", "line", "schema", "edge", "area", "polygon", "heatmap",
    "path"),
  asp = NULL
)
}
\arguments{
\item{g}{An object of class \code{g2r} or \code{g2Proxy} as returned by \code{\link[=g2]{g2()}} or
\code{\link[=g2_proxy]{g2_proxy()}}.}

\item{...}{Options to pass to the figure, including \code{\link[=asp]{asp()}},
and \code{\link[=adjust]{adjust()}}, \code{\link[=active]{active()}}, \code{\link[=selected]{selected()}}, and \code{\link[=config]{config()}},
other key value pairs of are passed to
\href{https://g2.antv.vision/en/docs/api/general/style/}{style}.}

\item{data}{A dataset (\code{data.frame} or \code{tibble}) to use to
draw the figure.}

\item{inherit_asp}{Whether to inherit the aspects paseed to
\code{\link[=g2]{g2()}} initialisation function.}

\item{sync}{Whether to sync the axis data (align) with that
used in other figures, set to \code{FALSE} to not sync or set to
a character string to use as name of sync group.}

\item{asp}{Aspects to override those passed to \code{...}.}
}
\description{
Add a path figure.
}
\keyword{internal}
