% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{fig_ribbon}
\alias{fig_ribbon}
\title{Ribbon}
\usage{
fig_ribbon(g2, ..., data = NULL, inherit_asp = TRUE, name = NULL)
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}

\item{...}{Any option, aesthetic (\code{\link{asp}}), or animation (\code{\link{Animation}}).}

\item{data}{A \code{data.frame} containing data to chart or a \code{list}.}

\item{inherit_asp}{Whether to inherit aspects from \code{g2r}.}

\item{name}{Name of figure, useful to apply functions to specific figures.}
}
\description{
Add a ribbon based ojn \code{ymin}, \code{ymax}, and optionally \code{group} aspects.
}
\examples{
df <- dplyr::tibble(
  x = 1:100,
  y = runif(100, 7, 18),
  y1 = runif(100, 5, 10),
  y2 = runif(100, 12, 20)
) \%>\% 
  dplyr::mutate(
    y = runif(dplyr::n(), y1 + 1, y2 - 1)
  )

g2(df, asp(x, y, ymin = y1, ymax = y2)) \%>\% 
  fig_line() \%>\% 
  fig_ribbon()

}
