% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{fig_rug}
\alias{fig_rug}
\title{Rug}
\usage{
fig_rug(g2, ..., x_axis = TRUE, y_axis = TRUE, opacity = 0.5,
  data = NULL, inherit_asp = TRUE, name = NULL)
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}

\item{...}{Any option, aesthetic (\code{\link{asp}}), or animation (\code{\link{Animation}}).}

\item{x_axis, y_axis}{Whether to show the ruf on the x and y axis.}

\item{opacity}{Opacity of individual ticks, passed to \code{\link{asp}}.}

\item{data}{A \code{data.frame} containing data to chart or a \code{list}.}

\item{inherit_asp}{Whether to inherit aspects from \code{g2r}.}

\item{name}{Name of figure, useful to apply functions to specific figures.}
}
\description{
Add rugs
}
\examples{
cars2 <- cars \%>\% 
  dplyr::bind_rows(cars) \%>\% 
  dplyr::mutate(
    speed = jitter(speed),
    dist = jitter(dist)
  )

g2(cars2, asp(speed, dist)) \%>\% 
  fig_point() \%>\% 
  fig_rug()

}
