% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{fig_smooth}
\alias{fig_smooth}
\title{Smooth}
\usage{
fig_smooth(
  g,
  ...,
  method = c("linear", "gaussian", "cosine", "epanechnikov", "quartic", "triangular",
    "tricube", "triweight", "uniform", "polynomial", "logarithmic", "boxcar", "power",
    "exponential"),
  band_width = 1,
  sync = TRUE,
  data = NULL,
  inherit_asp = TRUE
)
}
\arguments{
\item{g}{An object of class \code{g2r} or \code{g2Proxy} as returned by \code{\link[=g2]{g2()}} or
\code{\link[=g2_proxy]{g2_proxy()}}.}

\item{...}{Options to pass to the figure, including \code{\link[=asp]{asp()}},
and \code{\link[=adjust]{adjust()}}, \code{\link[=active]{active()}}, \code{\link[=selected]{selected()}}, and \code{\link[=config]{config()}},
other key value pairs of are passed to
\href{https://g2.antv.vision/en/docs/api/general/style/}{style}.}

\item{method}{Smoothing method to use.}

\item{band_width}{Step size for Silverman's algorithm.}

\item{sync}{Whether to sync the axis data (align) with that
used in other figures, set to \code{FALSE} to not sync or set to
a character string to use as name of sync group.}

\item{data}{A dataset (\code{data.frame} or \code{tibble}) to use to
draw the figure.}

\item{inherit_asp}{Whether to inherit the aspects paseed to
\code{\link[=g2]{g2()}} initialisation function.}
}
\description{
Add a smooth(ing) figure to the chart.
}
\details{
Requires the \code{x} and \code{y} aspects.

This is a convenience function for a quick smoothing, see
the online documentation to see how to use your own model
for more control.
}
\examples{
g2(cars, asp(speed, dist)) \%>\%
  fig_point() \%>\%
  fig_smooth(method = "gaussian")

g2(iris, asp(Sepal.Width, Sepal.Length, color = Species)) \%>\%
  fig_point() \%>\%
  fig_smooth()
}
