% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{fig_waffle}
\alias{fig_waffle}
\title{Waffle}
\usage{
fig_waffle(
  g,
  ...,
  n = 500,
  rows = 10,
  size = c(1, 1),
  gap = 0.1,
  min_size = 15,
  sync = TRUE,
  data = NULL,
  inherit_asp = TRUE,
  style = NULL
)
}
\arguments{
\item{g}{An object of class \code{g2r} or \code{g2Proxy} as returned by \code{\link[=g2]{g2()}} or
\code{\link[=g2_proxy]{g2_proxy()}}.}

\item{...}{Options to pass to the figure, including \code{\link[=asp]{asp()}},
and \code{\link[=adjust]{adjust()}}, \code{\link[=active]{active()}}, and \code{\link[=selected]{selected()}}.}

\item{n}{Number of squares to use.}

\item{rows}{Number of rows.}

\item{size}{Size of squares.}

\item{gap}{Gap between squares.}

\item{min_size}{Minimum size of squares.}

\item{sync}{Whether to sync the axis data (align) with that
used in other figures, set to \code{FALSE} to not sync or set to
a character string to use as name of sync group.}

\item{data}{A dataset (\code{data.frame} or \code{tibble}) to use to
draw the figure.}

\item{inherit_asp}{Whether to inherit the aspects paseed to
\code{\link[=g2]{g2()}} initialisation function.}

\item{style}{A list of options to pass to
\href{https://g2.antv.vision/en/docs/api/general/style}{style} that
customise the figure.}
}
\description{
Add a waffle figure to the chart.
}
\examples{
fruits <- data.frame(
 fruit = c("Apples", "Bananas", "Pears", "Oranges"),
 value = c(.45, .15, .35, .05) * 100
)

g2(fruits, asp(value, color = fruit)) \%>\% 
 fig_waffle() \%>\% 
 motif(padding = 50) \%>\% 
 axis_hide()

}
