% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{figures}
\alias{figures}
\alias{fig_interval}
\title{Interval}
\usage{
fig_interval(g, ..., sync = TRUE, data = NULL, inherit_asp = TRUE)
}
\arguments{
\item{g}{An object of class \code{g2r} as returned by \code{\link[=g2]{g2()}}.}

\item{...}{Options to pass to the figure, including \code{\link[=asp]{asp()}},
and \code{\link[=adjust]{adjust()}}.}

\item{sync}{Whether to sync the axis data (align).}

\item{data}{A dataset (\code{data.frame} or \code{tibble}) to use to
draw the figure.}

\item{inherit_asp}{Whether to inherit the aspects paseed to
\code{\link[=g2]{g2()}} initialisation function.}
}
\description{
Add an interval figure.
}
\examples{
g2(sleep, asp(ID, extra, color = group)) \%>\% 
 fig_interval()

df <- data.frame(
 cat = letters[1:5],
 value = c(0.15, .3, .65, .75, .9)
)

g2(df, asp(cat, value, color = cat, shape = "funnel")) \%>\% 
 fig_interval(adjust("symmetric")) \%>\% 
 coord_type("rect") \%>\% 
 coord_transpose() \%>\% 
 coord_scale(-1, 1) \%>\% 
 axis_hide()

}
