% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-options.R
\name{frame_var}
\alias{frame_var}
\title{Frame}
\usage{
frame_var(g2, var, ...)
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}

\item{var}{Bare column name of variable to frame.}

\item{...}{setting .}
}
\description{
Frame a variable.
}
\examples{
iris \%>\% 
  g2(asp(Sepal.Length, Sepal.Width)) \%>\% 
  fig_point(asp(color = Species)) \%>\% 
  frame_var(Sepal.Width, min = 0)

}
