% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fruits}
\alias{fruits}
\title{Fruits}
\format{A \code{tibble} of 2 columns and 4 rows.
\describe{
  \item{\code{fruit}}{ Fruit}
  \item{\code{value}}{ Percentage of fruit bought}
}}
\source{
Totally made up
}
\usage{
fruits
}
\description{
Fruits someone probably bought during a given week.
}
\keyword{datasets}
