% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2r.R
\name{g2}
\alias{g2}
\title{Initialise}
\usage{
g2(
  data = NULL,
  ...,
  width = NULL,
  height = NULL,
  elementId = NULL,
  digits = NULL,
  reorder = TRUE
)
}
\arguments{
\item{data}{A data.frame or tibble containing data to chart,
an object of class \code{igraph}, an object of class \code{ts},
or as \code{crosstalk::sharedDataset}.}

\item{...}{Aspects of the chart, see \code{\link[=asp]{asp()}}.}

\item{width, height}{Dimensions of the chart, accepts
any valid CSS unit e.g.: \verb{100\%}, numerics are treated
as pixels, e.g.: \code{400} = \verb{400px}.}

\item{elementId}{Valid CSS id attribute.}

\item{digits}{Maximum number of digits after the comma to show
on the chart.}

\item{reorder}{Whether to internally reorder the data, namely
the \code{x} and \code{color}. The \code{x} axis must be reordered in a descending
order for most data type since G2.js plots data as-is. Moreover,
\code{color} order of all data.frames passed either to this function or
subsequent \verb{fig_*} layers must be identical or the colors will
might match the legends on the plot. However, one may sometimes
not want the data to be reordered.}
}
\description{
Initialise a chart.
}
\examples{
g2(cars) \%>\%
  fig_point(asp(speed, dist))
}
