% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{g2_action}
\alias{g2_action}
\title{Plot Action}
\usage{
g2_action(plot_id, input_id, ..., data = NULL)
}
\arguments{
\item{plot_id}{Id of chart to interact with.}

\item{input_id}{Id of input that triggers the action.}

\item{...}{Aspects, see \code{\link[=asp]{asp()}}.}

\item{data}{Data.frame containing data to plot.}
}
\description{
Include dynamic elements in Rmarkdown.
}
