% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{g2_bar}
\alias{g2_bar}
\title{Bar}
\usage{
g2_bar(g2, ..., data = NULL)
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}

\item{...}{Any option or aesthetic.}

\item{data}{A \code{data.frame} containing data to chart.}
}
\description{
Bar
}
