% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.R
\name{g2_chart}
\alias{g2_chart}
\title{Charting Functions}
\usage{
g2_chart(g2, forcefit = "true", width = 500, height = 500,
  animate = TRUE, theme = "default", renderer = "canvas")
}
\description{
<Add Description>
}
\references{
\url{https://www.yuque.com/antv/g2-docs/api-chart}
}
