% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R
\name{g2_coord}
\alias{g2_coord}
\title{<Add Title>}
\usage{
g2_coord(g2, type = "rect", cfg = NULL, rotate = NULL,
  scale = NULL, reflect = NULL, transpose = NULL)
}
\arguments{
\item{coord}{The type of coordinate system.  Options are c("rect", "polar", "theta", "helix")}
}
\description{
<Add Description>
}
\references{
\url{https://www.yuque.com/antv/g2-docs/api-chart#ufv0rv}
}
