% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet.R
\name{g2_facet}
\alias{g2_facet}
\title{<Add Title>}
\usage{
g2_facet(g2, fields, type = "rect")
}
\arguments{
\item{type}{\itemize{
 \item{"rect"}
 \item{"list"}
 \item{"circle"}
 \item{"tree"}
 \item{"mirror"}
 \item{"matrix"}
}}
}
\description{
<Add Description>
}
