% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{g2_options}
\alias{g2_options}
\title{General Options}
\usage{
g2_options(g2, ...)
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}

\item{...}{Any options, see \href{https://www.yuque.com/antv/g2-docs-en/api-chart}{official documentation}.}
}
\description{
Set general chart options.
}
