% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{g2_proxy}
\alias{g2_proxy}
\title{Shiny Proxy}
\usage{
g2_proxy(id, ..., data = NULL, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{Id of chart to interact with.}

\item{...}{Aspects, see \code{\link[=asp]{asp()}}.}

\item{data}{Data.frame containing data to plot.}

\item{session}{A valid shiny session.}
}
\description{
Proxy to dynamically interact with the chart in shiny.
}
\examples{
library(shiny)

dataset <- data.frame(x = 1:100, y = runif(100, 1, 100))

ui <- fluidPage(
 g2rOutput("plot"),
 actionButton("add", "Add figure")
)

server <- function(input, output, session){

 output$plot <- renderG2r({
   g2(dataset, asp(x, y)) \%>\% 
     fig_point()
 })

 observeEvent(input$add, {
   df <- data.frame(x = 1:100, y = runif(100, 1, 100))
   g2_proxy("plot", data = df) \%>\% 
     fig_point(asp(x, y)) \%>\% 
     render()
   })

}

if(interactive())
 shinyApp(ui, server)

}
