% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{g2_scale}
\alias{g2_scale}
\title{<Add Title>}
\usage{
g2_scale(g2, xtype, ytype, xopts, yopts)
}
\arguments{
\item{xtype, ytype}{\itemize
\item{"identity"}
\item{"linear"}
\item{"cat"}
\item{"time"}
\item{"timeCat"}
\item{"log"}
\item{"pow"}}
}
\description{
<Add Title>
}
