% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{g2_scale_label}
\alias{g2_scale_label}
\title{Scale label}
\usage{
g2_scale_label(g2, label)
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}

\item{label}{A configuration list or a JavaScript funtion.}
}
\description{
Scale label.
}
