% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{g2_scale_opacity}
\alias{g2_scale_opacity}
\title{Scale opacity}
\usage{
g2_scale_opacity(g2, opacity)
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}

\item{opacity}{A JavaScript function.}
}
\description{
Scale opacity.
}
