% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{g2_scale_size}
\alias{g2_scale_size}
\title{Scale size}
\usage{
g2_scale_size(g2, size)
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}

\item{size}{A vector of sizes (\code{min, max}) or a JavaScript function.}
}
\description{
Scale size.
}
