% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{g2_scale_tooltip}
\alias{g2_scale_tooltip}
\title{Scale tooltip}
\usage{
g2_scale_tooltip(g2, tooltip)
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}

\item{tooltip}{A JavaScript function.}
}
\description{
Scale tooltip.
}
