% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animation.R
\docType{data}
\name{G2animation}
\alias{G2animation}
\alias{Animation}
\alias{new_animation}
\title{Animations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Animation

new_animation()
}
\description{
Helper function to build animations.
}
\examples{
# create animation
anim <- Animation$
  new()$
  appear(duration = 2000, delay = 500)

iris \%>\% 
  g2(asp(Sepal.Length, Sepal.Width, color = Species)) \%>\% 
  fig_point(anim) # pass animation to the geom

}
\keyword{datasets}
