% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2plot.R
\name{g2plot}
\alias{g2plot}
\title{g2plot main function}
\usage{
g2plot(data, width = NULL, height = NULL, elementId = NULL, ...)
}
\arguments{
\item{data}{data you will be plot}
}
\description{
<Add Description>
}
