% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2R.R
\name{g2R-shiny}
\alias{g2R-shiny}
\alias{g2ROutput}
\alias{renderG2R}
\title{Shiny bindings for g2R}
\usage{
g2ROutput(outputId, width = "100\%", height = "400px")

renderG2R(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a g2R}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using g2R within Shiny
applications and interactive Rmd documents.
}
