% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2.R
\name{g2r}
\alias{g2r}
\alias{g2}
\title{Initialise}
\usage{
g2(data = NULL, plan = NULL, ..., width = NULL, height = NULL,
  elementId = NULL)
}
\arguments{
\item{data}{A \code{data.frame} containing data to chart.}

\item{plan}{Mapping aesthetics as returned by \code{\link{plan}}.}

\item{...}{Any general options.}

\item{width, height}{Dimensions of chart.}

\item{elementId}{ID of \code{DOM} container.}
}
\description{
Initialise a \code{g2r} chart.
}
\examples{
cars \%>\% 
  g2(plan(speed, dist)) \%>\% 
  fig_point()

}
