% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{g2r_globals}
\alias{g2r_globals}
\title{Globals}
\usage{
g2r_globals(..., font = NULL, renderer = NULL)
}
\arguments{
\item{...}{Passed to \code{\link{custom_style}}}

\item{font}{Passed to \code{\link{set_font}}}

\item{renderer}{Passed to \code{\link{set_renderer}}}
}
\description{
Set global options, all charts in session will use these options.
}
