% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauge.R
\name{gauge2_}
\alias{gauge2_}
\title{Gauge2}
\usage{
gauge2_(g, handler, fn)
}
\arguments{
\item{g}{An object of class \code{g2r} as returned by \code{\link[=g2]{g2()}}.}

\item{handler}{Handler to pass as second argument to
the JavaScript function.}

\item{fn}{Name of the function.}
}
\description{
While \code{\link[=gauge_]{gauge_()}} applies the \code{scale} to the
global \code{chart} object, \code{\link[=gauge2_]{gauge2_()}} applies it to the
\code{view}.
}
\keyword{internal}
