% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{gauge_axis}
\alias{gauge_axis}
\title{Axis}
\usage{
gauge_axis(g2, var, ...)
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}

\item{var}{Variable to scale.}

\item{...}{Configuration options or a logical indicating whether to show the axis.}
}
\description{
Scale and modify axis.
}
\examples{
g2(cars, plan(speed, dist)) \%>\% 
  fig_point() \%>\% 
  gauge_axis(speed, FALSE) 

}
