% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauge.R
\name{gauge_opacity}
\alias{gauge_opacity}
\title{Gauge opacity}
\usage{
gauge_opacity(g2, opacity)
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}

\item{opacity}{A JavaScript function.}
}
\description{
Gauge opacity.
}
