% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauge.R
\name{gauge_shape}
\alias{gauge_shape}
\title{Gauge shape}
\usage{
gauge_shape(g2, shapes = NULL, callback = NULL)
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}

\item{shapes}{A vector of shapes.}

\item{callback}{A JavaScript callback function (see \code{\link{cb}}) which returns a color.}
}
\description{
Gauge shapes.
}
