% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauge.R
\name{gauge_size}
\alias{gauge_size}
\title{Gauge size}
\usage{
gauge_size(g2, range = NULL, callback = NULL)
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}

\item{range}{A vector indicating the minimum and maximum sizes.}

\item{callback}{A JavaScript callback function (see \code{\link{cb}}) which returns a color.}
}
\description{
Gauge size given a range or a JavaScript callback function.
}
\examples{
func <- cb("function(dist){
  if(dist > 60)
    return 20
  return 5
}")

g2(cars, asp(speed, dist, size = dist)) \%>\% 
  fig_point() \%>\% 
  gauge_size(callback = func)

}
