% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauge.R
\name{gauge_style}
\alias{gauge_style}
\title{Gauge label}
\usage{
gauge_style(g2, ...)
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}

\item{...}{Any option, aesthetic (\code{\link{asp}}), or animation (\code{\link{Animation}}).}
}
\description{
Gauge label.
}
\examples{
g2(mtcars, asp(mpg, qsec, shape = drat)) \%>\% 
  fig_point() 

}
