% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauge.R
\name{gaugeViews}
\alias{gaugeViews}
\alias{gauge_color}
\alias{gauge_size}
\alias{gauge_shape}
\alias{gauge_tooltip}
\alias{gauge_label}
\alias{gauge_style}
\alias{gauge_interplay}
\title{Gauge Aspects}
\usage{
gauge_color(g, ...)

gauge_size(g, ...)

gauge_shape(g, ...)

gauge_tooltip(g, ...)

gauge_label(g, ...)

gauge_style(g, ...)

gauge_interplay(g, ...)
}
\arguments{
\item{g}{An object of class \code{g2r} or \code{g2Proxy} as returned by \code{\link[=g2]{g2()}} or
\code{\link[=g2_proxy]{g2_proxy()}}.}

\item{...}{Arguments to customise the gauge.
Generally, key value pairs of options, a vector of hex colors,
or a JavaScript function (wrapped in \code{\link[htmlwidgets:JS]{htmlwidgets::JS()}}).}
}
\description{
Customise aspects of the chart.
}
\examples{
# base plot
g <- g2(cars, asp(speed, dist)) \%>\%
  fig_point(asp(color = speed))

# color with vector
g \%>\% gauge_color(c("red", "white", "blue"))

# color with callback
cb <- "function(speed){
 if(speed > 10){
   return 'blue';
 }
 return 'red';
}"

g \%>\% gauge_color(htmlwidgets::JS(cb))
}
\seealso{
\link{gauge} to gauge aspects of the grid and axis.
}
