% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gaus}
\alias{gaus}
\title{Gaus}
\format{A \code{tibble} of 2 columns and 2000 rows.
\describe{
  \item{\code{x}}{ X Axis}
  \item{\code{y}}{ Y Axis}
}}
\source{
Somewhere mysterious
}
\usage{
gaus
}
\description{
Gaussian distribution.
}
\keyword{datasets}
