% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{geoms}
\alias{geoms}
\alias{fig_interval}
\alias{fig_line}
\alias{fig_point}
\alias{fig_path}
\alias{fig_area}
\alias{fig_polygon}
\alias{fig_schema}
\alias{fig_edge}
\alias{fig_heatmap}
\alias{fig_point_jitter}
\alias{fig_point_stack}
\alias{fig_point_dodge}
\alias{fig_interval_stack}
\alias{fig_interval_dodge}
\alias{fig_interval_symmetric}
\alias{fig_area_stack}
\alias{fig_schema_dodge}
\title{Firgures}
\usage{
fig_interval(g2, ..., data = NULL, inherit_gaes = TRUE)

fig_line(g2, ..., data = NULL, inherit_gaes = TRUE)

fig_point(g2, ..., data = NULL, inherit_gaes = TRUE)

fig_path(g2, ..., data = NULL, inherit_gaes = TRUE)

fig_area(g2, ..., data = NULL, inherit_gaes = TRUE)

fig_polygon(g2, ..., data = NULL, inherit_gaes = TRUE)

fig_schema(g2, ..., data = NULL, inherit_gaes = TRUE)

fig_edge(g2, ..., data = NULL, inherit_gaes = TRUE)

fig_heatmap(g2, ..., data = NULL, inherit_gaes = TRUE)

fig_point_jitter(g2, ..., data = NULL, inherit_gaes = TRUE)

fig_point_stack(g2, ..., data = NULL, inherit_gaes = TRUE)

fig_point_dodge(g2, ..., data = NULL, inherit_gaes = TRUE)

fig_interval_stack(g2, ..., data = NULL, inherit_gaes = TRUE)

fig_interval_dodge(g2, ..., data = NULL, inherit_gaes = TRUE)

fig_interval_symmetric(g2, ..., data = NULL, inherit_gaes = TRUE)

fig_area_stack(g2, ..., data = NULL, inherit_gaes = TRUE)

fig_schema_dodge(g2, ..., data = NULL, inherit_gaes = TRUE)
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}

\item{...}{Any option, aesthetic (\code{\link{gaes}}), or animation (\code{\link{Animation}}).}

\item{data}{A \code{data.frame} containing data to chart.}

\item{inherit_gaes}{Whether to inherit aesthetics from \code{g2r}.}
}
\description{
Figures, equivalent to \code{ggplot2}'s geometries.
}
\examples{
iris \%>\% 
  g2(plan(Sepal.Length, Sepal.Width, color = Species)) \%>\% 
  fig_point()

}
