% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.R
\name{get_adjust}
\alias{get_adjust}
\title{Get Adjust}
\usage{
get_adjust(...)
}
\arguments{
\item{...}{Any argument.}
}
\description{
Get any \code{\link[=adjust]{adjust()}} from the three dot construct.
}
\examples{
\dontrun{
foo <- function(...){
 get_adjust(...)
}

foo(1)
foo(adjust("stack"))
}

}
\keyword{internal}
