% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspects.R
\name{get_asp}
\alias{get_asp}
\title{Get Aspects}
\usage{
get_asp(...)
}
\arguments{
\item{...}{Three dot passed from parent from which to \emph{retrieve} aspects.}
}
\description{
Get aspects from three dot construct
}
\examples{
foo <- function(...){
 get_asp(...)
}

foo(asp(speed))

}
\keyword{internal}
