% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{get_axis_column_names}
\alias{get_axis_column_names}
\title{Retrieve Column Names}
\usage{
get_axis_column_names(g, axis = c("x", "y"))
}
\arguments{
\item{g}{An object of class \code{g2r} as returned by \code{\link[=g2]{g2()}}.}

\item{axis}{Axis to retrieve the used column names.}
}
\description{
Retrieve column names used for position (x, and y).
}
\keyword{internal}
