% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspects.R
\name{get_combined_asp}
\alias{get_combined_asp}
\title{Combined Aspects}
\usage{
get_combined_asp(g, ..., inherit_asp = FALSE)
}
\arguments{
\item{g}{An object of class \code{g2r} or \code{g2Proxy} as returned by \code{\link[=g2]{g2()}} or
\code{\link[=g2_proxy]{g2_proxy()}}.}

\item{...}{Three dots from parents (containing aspects).}

\item{inherit_asp}{Whether to inherit the aspects paseed to
\code{\link[=g2]{g2()}} initialisation function.}

\item{inherits_asp}{Whether the main aspects are inherited.}
}
\description{
Retrieve the combined main and figure-level aspects.
}
\keyword{internal}
