% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_config}
\alias{get_config}
\title{Get Config}
\usage{
get_config(...)
}
\arguments{
\item{...}{Three dot passed from parent from which
to \emph{retrieve} config.}
}
\description{
Get config from three dot construct
}
\examples{
\dontrun{
foo <- function(...){
 get_conf(...)
}

foo(conf("hello"))
}

}
\keyword{internal}
