% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{get_elements}
\alias{get_elements}
\title{Get or Remove Elements}
\usage{
get_elements(...)
}
\arguments{
\item{...}{Passed from \code{\link[=motif]{motif()}}.}
}
\description{
Get or remove elements from three dot constructs.
}
\keyword{internal}
