% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{get_global_chart_opts}
\alias{get_global_chart_opts}
\title{Get chart options}
\usage{
get_global_chart_opts()
}
\description{
We don't currently provide a way to set the theme
globally but it'll be useful then.
}
\keyword{internal}
