% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{get_map_data}
\alias{get_map_data}
\title{Get Map}
\usage{
get_map_data(name = "World")
}
\arguments{
\item{name}{The name of the map.}
}
\description{
Retrieve map data from \href{https://github.com/JohnCoene/echarts4r.maps}{echarts4r.maps}.
}
