% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{get_state}
\alias{get_state}
\title{Get States}
\usage{
get_state(...)
}
\arguments{
\item{...}{Three dot passed from parent from which
to \emph{retrieve} states.}
}
\description{
Get states from three dot construct
}
\examples{
\dontrun{
foo <- function(...){
 get_state(...)
}

foo(active(stroke = "red"))
}

}
\keyword{internal}
