% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltip.R
\name{get_tooltip}
\alias{get_tooltip}
\title{Get Tooltip}
\usage{
get_tooltip(...)
}
\arguments{
\item{...}{Any argument.}
}
\description{
Get any \code{\link[=tooltip]{tooltip()}} from the three dot construct.
}
\examples{
\dontrun{
foo <- function(...){
 get_tooltip(...)
}

foo(1)
foo(tooltip("callback", x, y))
}

}
\keyword{internal}
