% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspects.R
\name{has_asp}
\alias{has_asp}
\title{Has Aspects}
\usage{
has_asp(asp)
}
\description{
Check whether the list contains aspects. This is
because functions to select or retrieve aspects
return lists of length 0 when none are found.
}
\examples{
\dontrun{
has_asp(asp())
has_asp(asp(dist))
}

}
\keyword{internal}
