% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{igraph_to_list}
\alias{igraph_to_list}
\title{igraph to List}
\usage{
igraph_to_list(data)
}
\arguments{
\item{data}{An igraph object.}
}
\value{
\code{list} of length 2.
}
\description{
Returns a list of nodes and edges.
}
\keyword{internal}
