% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{info}
\alias{info}
\alias{info_text}
\alias{info_image}
\alias{info_arc}
\alias{info_line}
\alias{info_vline}
\alias{info_hline}
\alias{info_region}
\alias{info_region_filter}
\alias{info_marker}
\alias{info_data_region}
\alias{info_shape}
\alias{info_html}
\title{Text}
\usage{
info_text(g, ..., style = NULL, data = NULL)

info_image(g, ..., style = NULL, data = NULL)

info_arc(g, ..., style = NULL, data = NULL)

info_line(g, ..., style = NULL, data = NULL)

info_vline(g, ..., style = NULL, data = NULL)

info_hline(g, ..., style = NULL, data = NULL)

info_region(g, ..., style = NULL, data = NULL)

info_region_filter(g, ..., style = NULL, data = NULL)

info_marker(g, ..., style = NULL, data = NULL)

info_data_region(g, ..., style = NULL, data = NULL)

info_shape(g, ..., style = NULL, data = NULL)

info_html(g, ..., style = NULL, data = NULL)
}
\arguments{
\item{g}{An object of class \code{g2r} as returned by \code{\link[=g2]{g2()}}.}

\item{...}{Options to pass to the informational annotation.}

\item{style}{A list of options defning the style.}

\item{data}{A dataset to use with \code{\link[=asp]{asp()}}.}
}
\description{
Add a point figure.
}
\details{
\code{info_vline}, and \code{info_hline} use the \code{x}, and \code{y}
\code{\link[=asp]{asp()}} for placement.
}
\examples{
df <- head(cars, 5)

g2(cars, asp(speed, dist)) \%>\%
 fig_point() \%>\% 
 info_text(
   position = c(20, 35), content = "Look here!"
 ) \%>\% 
 info_text(
   asp(speed, dist),
   content = "Using aspects",
   data = df
 )

g2(cars, asp(speed, dist)) \%>\%
 fig_point() \%>\% 
 info_vline(asp(x = 20)) \%>\% 
 info_hline(asp(y = 20))

}
\seealso{
\href{https://antv-g2.gitee.io/en/docs/api/general/annotation}{Official annotation documentation}
for defails pon what to pass to \code{...}, and \code{\link[=asp]{asp()}}.
}
