% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{info_aspects_data}
\alias{info_aspects_data}
\title{Select Aspects from Data}
\usage{
info_aspects_data(asp, data = NULL)
}
\arguments{
\item{asp}{Aspects as returned by \code{\link[=asp]{asp()}}.}

\item{data}{Data.frame.}
}
\description{
Select aspects from the data for info.
}
