% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{info_primitive}
\alias{info_primitive}
\title{Info}
\usage{
info_primitive(
  g,
  ...,
  type = c("text", "image", "arc", "line", "region", "regionFilter", "dataMarker",
    "dataRegion", "shape", "html"),
  style = NULL,
  data = NULL,
  asp = NULL
)
}
\arguments{
\item{g}{An object of class \code{g2r} as returned by \code{\link[=g2]{g2()}}.}

\item{...}{Options to pass to the figure, including \code{\link[=asp]{asp()}},
and \code{\link[=adjust]{adjust()}}, \code{\link[=active]{active()}}, and \code{\link[=selected]{selected()}}.}

\item{type}{Type of annotation.}

\item{style}{Style list to pass to the annotation.}

\item{data}{Optional data.frame to use with the \code{\link[=asp]{asp()}}.}

\item{asp}{Aspects, if \code{NULL} then the function retrieves them
from the \code{...}.}
}
\description{
Primitive to create informational annotations.
}
\keyword{internal}
