% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{input_button}
\alias{input_button}
\title{Button Input}
\usage{
input_button(id, label, class = "default")
}
\arguments{
\item{id}{Id of the button.}

\item{label}{Label to display.}

\item{class}{Class of the button.}
}
\description{
Add a button input.
}
\details{
The \code{class} argument defines the style of
the button in Bootstrap 3, generally accepts:\code{for}
\itemize{
\item \code{default}
\item \code{info}
\item \code{success}
\item \code{warning}
\item \code{danger}
}
}
