% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{input_checkbox}
\alias{input_checkbox}
\title{Checkbox and Radio}
\usage{
input_checkbox(id, label, choices, selected = NULL, inline = TRUE)
}
\arguments{
\item{id}{Id of input.}

\item{label}{Label of the input.}

\item{choices}{Vector of choices to define
either the checboxes or radio inputs.}

\item{selected}{Vector of \code{choices} that are
selected by default.}

\item{inline}{Whether the input is inline.}
}
\description{
Add a checkbox or radio input.
}
