% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{input_select}
\alias{input_select}
\title{Select Input}
\usage{
input_select(id, label, choices)
}
\arguments{
\item{id}{Valid CSS id of the element.}

\item{label}{Label to display.}

\item{choices}{Vector of choices}
}
\description{
Select Input
}
