% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{input_slider}
\alias{input_slider}
\title{Slider Input}
\usage{
input_slider(id, label, value, min, max, step = 1)
}
\arguments{
\item{id}{Valid CSS id of the element.}

\item{label}{Label to display.}

\item{value}{Initial value of the slider.}

\item{min, max}{Minimum and maximum value the slider can be set.}

\item{step}{Interval between steps.}
}
\description{
Add a slider to an R markdown document.
}
\examples{
input_slider(
 "mySlider",
 "The label",
 value = 5,
 min = 0,
 max = 10
)

}
