% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact.R
\name{interact}
\alias{interact}
\alias{interact_brush}
\alias{interact_slider}
\alias{interact_zoom}
\alias{interact_drag}
\title{Brush}
\usage{
interact_brush(g2)

interact_slider(g2, ...)

interact_zoom(g2, ...)

interact_drag(g2, ...)
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}

\item{...}{Any option, aesthetic (\code{\link{asp}}), or animation (\code{\link{Animation}}).}
}
\description{
Add brush interactions.
}
\examples{
range <- range(cars$speed)

g2(cars, asp(speed, dist)) \%>\% 
  fig_point() \%>\% 
  interact_brush()

}
