% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{interaction}
\alias{interaction}
\title{Interactions}
\usage{
interaction(g, ...)
}
\arguments{
\item{g}{An object of class \code{g2r} as returned by \code{\link[=g2]{g2()}}.}

\item{...}{String(s) defining interactions.}
}
\description{
Add interactions to the chart.
}
\examples{
df <- data.frame(
 x = letters,
 y = runif(26)
)

g2(df, asp(x, y)) \%>\%
 fig_interval(
   selected(fill = "orange")
 ) \%>\% 
 interaction("element", "selected")

}
