% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interplay.R
\name{interplay}
\alias{interplay}
\alias{conf_interplay}
\alias{remove_interplay}
\title{Interplay}
\usage{
conf_interplay(g, ...)

remove_interplay(g, ...)
}
\arguments{
\item{g}{An object of class \code{g2r} or \code{g2Proxy} as returned by \code{\link[=g2]{g2()}} or
\code{\link[=g2_proxy]{g2_proxy()}}.}

\item{...}{String(s) defining interactions.}
}
\description{
Configure global interplay (interactions) for the chart.
See \code{\link[=gauge_interplay]{gauge_interplay()}} to customise figure-level interplay.
}
\examples{
df <- data.frame(
 x = letters,
 y = runif(26)
)

g2(df, asp(x, y)) \%>\%
 fig_interval(
   selected(fill = "orange")
 ) \%>\% 
 conf_interplay("element", "selected")

}
