% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.R
\name{is_adjust}
\alias{is_adjust}
\title{Adjust Check}
\usage{
is_adjust(x)
}
\arguments{
\item{x}{Object to check.}
}
\value{
A boolean.
}
\description{
Checks whether the object is of class \code{adjust},
as returned by \code{\link[=adjust]{adjust()}}.
}
\examples{
\dontrun{
is_adjust(1)
is_adjust(adj("stack"))
}

}
\keyword{internal}
