% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspects.R
\name{is_asp}
\alias{is_asp}
\title{Aspect Check}
\usage{
is_asp(x)
}
\arguments{
\item{x}{Object to check.}
}
\value{
A boolean.
}
\description{
Checks whether the object is of class \code{aspects},
as returned by \code{\link[=asp]{asp()}}.
}
\examples{
\dontrun{
is_asp(1)
is_asp(asp(dist))
}

}
\keyword{internal}
