% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{is_config}
\alias{is_config}
\title{Config Check}
\usage{
is_config(x)
}
\arguments{
\item{x}{Object to check.}
}
\value{
A boolean.
}
\description{
Checks whether the object is of class \code{config},
as returned by \code{\link[=config]{config()}}.
}
\examples{
\dontrun{
is_config(1)
is_config(conf("hello"))
}

}
\keyword{internal}
