% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{is_element}
\alias{is_element}
\title{Element Check}
\usage{
is_element(obj)
}
\arguments{
\item{obj}{An object to check.}
}
\value{
If \code{\link[=element]{element()}} returns \code{TRUE},
otherwise returns \code{FALSE}.
}
\description{
Check whether the object is an \code{\link[=element]{element()}}.
}
\keyword{internal}
