% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{is_state}
\alias{is_state}
\title{State Check}
\usage{
is_state(x)
}
\arguments{
\item{x}{Object to check.}
}
\value{
A boolean.
}
\description{
Checks whether the object is of class \code{state},
as returned by \code{\link[=selected]{selected()}} and \code{\link[=active]{active()}}.
}
\examples{
\dontrun{
is_state(1)
is_state(selected(fill = "red"))
}

}
\keyword{internal}
