% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltip.R
\name{is_tooltip}
\alias{is_tooltip}
\title{Tooltip Check}
\usage{
is_tooltip(x)
}
\arguments{
\item{x}{Object to check.}
}
\value{
A boolean.
}
\description{
Checks whether the object is of class \code{tooltip},
as returned by \code{\link[=tooltip]{tooltip()}}.
}
\examples{
\dontrun{
is_tooltip(1)
is_tooltip(tooltip("callback", x, y))
}

}
\keyword{internal}
