% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{layout_igraph}
\alias{layout_igraph}
\title{Layout with igraph}
\usage{
layout_igraph(g, ..., method = igraph::layout_nicely)
}
\arguments{
\item{g}{An object of class \code{g2r} or \code{g2Proxy} as returned by \code{\link[=g2]{g2()}} or
\code{\link[=g2_proxy]{g2_proxy()}}.}

\item{...}{Any options to pass to the \code{method} function.}

\item{method}{An igraph layout function to compute the
nodes and edges (source and target) position on the canvas.}
}
\description{
Layout the graph using an igraph layout function.
This function only works with the graph was initialised
with an object of class \code{igraph}.
}
\details{
The function runs the \code{method} to obtain the \code{x}
and \code{y} coordinates. These are added to the \code{nodes}
data.frame (extracted from initial graph) and to the edges
data.frame, as \code{x} and \code{y} \emph{nested} columns, e.g.:
\code{c(source_x, target_x)}. These \code{x} and \code{y} coordinates
can be used in \code{\link[=asp]{asp()}} (see example).
}
\examples{
ig <- igraph::make_ring(100)

# use x and y for positioning
g2(ig, asp(x, y)) \%>\%
 layout_igraph() \%>\% 
 fig_edge() \%>\% 
 fig_point(asp(shape = "circle")) \%>\% 
 axis_hide()

}
