% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{legend}
\alias{legend}
\alias{conf_legend}
\alias{leg}
\title{Legend}
\usage{
conf_legend(g2, var, ...)

leg(var, ...)
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}

\item{var}{Bare column name of legend to apply changes to.}

\item{...}{Configuration options or a logical indicating whether to show the legend.}
}
\description{
Configure the chart's legend.
}
\examples{
g <- g2(cars, plan(speed, dist, color = dist)) \%>\% 
  fig_point()

g \%>\% conf_legend(dist, FALSE)

g \%>\% conf_legend(dist, position = "right")

}
