% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{make_geoms}
\alias{make_geoms}
\title{Make geoms}
\usage{
make_geoms(theme, geoms)
}
\arguments{
\item{theme}{Base theme object.}

\item{geoms}{List of geoms as returned by \code{\link[=get_elements]{get_elements()}}.}
}
\description{
Make geometries structure for theme.
}
\keyword{internal}
