% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_data}
\alias{map_data}
\alias{get_gadm_data}
\alias{get_map_data}
\alias{get_world_map}
\title{Get Map}
\usage{
get_gadm_data(iso3c, level = c(0, 1, 2, 3, 4), keep = 0.05)

get_map_data(region = ".", level = c("region", "subregion"), name = "world")

get_world_map()
}
\arguments{
\item{iso3c}{Iso3c code of the country to retrieve, e.g.: \code{USA}.}

\item{level}{Level of the polygons to draw, either the region or
subregion.}

\item{keep}{Proportion of points to retain, it is highly recommended
to reduce the detail of the map or it will take too long to load in the
browser. Set to \code{NULL} to not reduce the amount of details.}

\item{region}{Character vector that names the polygons to draw.}

\item{name}{Name of the database to use.}
}
\description{
Retrieve map data to pass to the \code{maps}
argument of the \code{fig_map()} function.
}
\section{Functions}{

\itemize{
\item \code{get_gadm_data}: Retrieves country-level data from \href{https://gadm.org/}{gadm.org}.
\item \code{get_map_data}: Uses the \code{maps::map()} function to retrieve the map data, similar to \code{ggplot2::map_data}.
\item \code{get_world_map}: Retrives a world map (from GeoJSON).
}
}

