% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animation.R
\name{new_animation}
\alias{new_animation}
\title{New Animation}
\usage{
new_animation()
}
\description{
Convenience function to create a new animation,
equivalent to \code{Animation$new()}.
}
\seealso{
\link{Animation}
}
