% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{order_data}
\alias{order_data}
\title{Order Data}
\usage{
order_data(data, cols)
}
\arguments{
\item{data}{Data to reorder.}

\item{cols}{Columns, only uses the first which is assumes
is the \code{x} axis.}
}
\description{
Fortunately or unfortunately G2.js does not reorder
the data and uses it as-is.
}
\details{
This function reorders along the \code{x} axis to make sure
most plots look correct, see the \code{NO_REORDER_TYPES} for
figures types that are excluded from reordering.
}
\keyword{internal}
