% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauge.R
\name{palettes}
\alias{palettes}
\alias{gauge_color_viridis}
\alias{gauge_color_plasma}
\alias{gauge_color_inferno}
\alias{gauge_color_magma}
\alias{gauge_color_accent}
\alias{gauge_color_dark2}
\alias{gauge_color_paired}
\alias{gauge_color_pastel1}
\alias{gauge_color_pastel2}
\alias{gauge_color_set1}
\alias{gauge_color_set2}
\alias{gauge_color_set3}
\alias{gauge_color_neon}
\alias{gauge_color_std}
\alias{gauge_color_pink}
\alias{gauge_color_orange}
\alias{gauge_color_lime}
\alias{gauge_color_aw}
\title{Color Palettes}
\usage{
gauge_color_viridis(g)

gauge_color_plasma(g)

gauge_color_inferno(g)

gauge_color_magma(g)

gauge_color_accent(g)

gauge_color_dark2(g)

gauge_color_paired(g)

gauge_color_pastel1(g)

gauge_color_pastel2(g)

gauge_color_set1(g)

gauge_color_set2(g)

gauge_color_set3(g)

gauge_color_neon(g)

gauge_color_std(g)

gauge_color_pink(g)

gauge_color_orange(g)

gauge_color_lime(g)

gauge_color_aw(g)
}
\arguments{
\item{g}{An object of class \code{g2r} or \code{g2Proxy} as returned by \code{\link[=g2]{g2()}} or
\code{\link[=g2_proxy]{g2_proxy()}}.}
}
\description{
Convenience function to easily apply colors palettes.
}
\section{Palettes}{


Palletes from the \code{viridisLite} package, ideal for:
continuous data.
\itemize{
\item \code{gauge_color_viridis}
\item \code{gauge_color_plasma}
\item \code{gauge_color_inferno}
\item \code{gauge_color_magma}
}

Qualtitative palettes from color brewer:
\itemize{
\item \code{gauge_color_accent}
\item \code{gauge_color_dark2}
\item \code{gauge_color_paired}
\item \code{gauge_color_pastel1}
\item \code{gauge_color_pastel2}
\item \code{gauge_color_set1}
\item \code{gauge_color_set2}
\item \code{gauge_color_set3}
}

Palettes taken from \href{https://coolors.co}{coolors.co}:
\itemize{
\item \code{gauge_color_neon}
\item \code{gauge_color_std}
\item \code{gauge_color_orange}
\item \code{gauge_color_pink}
\item \code{gauge_color_lime}
}

Palettes from \code{awtools} package:
\itemize{
\item \code{gauge_color_aw}
}
}

