% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facets.R
\name{plane_wrap}
\alias{plane_wrap}
\title{Plane}
\usage{
plane_wrap(g2, facets, type = c("rect", "list", "circle", "tree",
  "mirror", "matrix"), ...)
}
\description{
Plane wrap is similar to the \code{facet_*} family of functions from the \code{ggplot2} package.
}
