% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{plot_callbacks}
\alias{plot_callbacks}
\title{Plot Callbacks}
\usage{
plot_callbacks(g2)
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}
}
\value{
A \code{list} containing \code{x} and \code{y} pixel coordinates of clicked/entered/left.
}
\description{
Add basic plot callbacks, will only be available in Shiny.
}
\details{
This function will add the following callbacks
\itemize{
  \item{\code{plot_enter}}
  \item{\code{plot_leave}}
  \item{\code{plot_click}}  
}
}
