% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preRender.R
\name{pre_render}
\alias{pre_render}
\title{Pre Render Hook}
\usage{
pre_render(g)
}
\arguments{
\item{g}{g2r obeject as returned by \code{\link[=g2]{g2()}}.}
}
\value{
A modified version of \code{g}.
}
\description{
Pre Render Hook
}
\keyword{internal}
