% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{proxy_data}
\alias{proxy_data}
\title{Serialise}
\usage{
proxy_data(data = NULL)
}
\arguments{
\item{data}{A dataframe or \code{NULL}.}
}
\description{
Serialise the data from a data frame to a rowwise
list. This is because proxies do not come with a
serialiser that can be adapted.
}
