% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick.R
\name{qg2}
\alias{qg2}
\title{Quick Plot}
\usage{
qg2(object, ..., conf_level = 0.95, intercept = FALSE, names = NULL)
}
\arguments{
\item{object}{An object containing data to plot,
often a model.}

\item{...}{Ignored}

\item{conf_level}{Confidence level.}

\item{intercept}{Whether to display the intercept.}

\item{names}{Names of the models.}
}
\description{
Draw a quick plot.
}
