% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{remove_figure}
\alias{remove_figure}
\title{Remove a figure}
\usage{
remove_figure(g, index)
}
\arguments{
\item{g}{An object of class \code{g2r} or \code{g2Proxy} as returned by \code{\link[=g2]{g2()}} or
\code{\link[=g2_proxy]{g2_proxy()}}.}

\item{index}{Index of figure to remove.}
}
\description{
Remove a figure from the plot.
}
\examples{
library(shiny)

df <- data.frame(
 x = 1:100,
 y = runif(100),
 z = runif(100)
)

ui <- fluidPage(
 g2rOutput("plot"),
 actionButton("rm", "Randomly remove a figure")
)

server <- function(input, output){

 output$plot <- renderG2r({
   g2(df, asp(x, y)) \%>\% 
     fig_point() \%>\% 
     fig_line(asp(y = z))
 })

 observeEvent(input$rm, {
   g2_proxy("plot") \%>\% 
     remove_figure(sample(1:2, 1))
 })

}

if(interactive())
 shinyApp(ui, server)

}
