% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{render}
\alias{render}
\title{Render}
\usage{
render(g, update = TRUE)
}
\arguments{
\item{g}{An object of class \code{g2Proxy} as returned
by \code{\link[=g2_proxy]{g2_proxy()}}.}

\item{update}{Whether to trigger the update process.}
}
\description{
Render proxy calls.
}
