% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauge.R
\name{scale-color}
\alias{scale-color}
\alias{gauge_colour}
\alias{gauge_color}
\title{Gauge colour}
\usage{
gauge_colour(g2, colors)

gauge_color(g2, colors)
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}

\item{colors}{A vector of colors or a JavaScript function.}
}
\description{
Gauge colour, similar to the \code{scale_colour_*} family of functions from the \code{ggplot2} package.
}
