% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{scrollbar}
\alias{scrollbar}
\title{Scrollbar}
\usage{
scrollbar(g, ...)
}
\arguments{
\item{g}{An object of class \code{g2r} or \code{g2Proxy} as returned by \code{\link[=g2]{g2()}} or
\code{\link[=g2_proxy]{g2_proxy()}}.}

\item{...}{Options to pass to the slider.}
}
\description{
Add a scrollbar to the chart.
}
\seealso{
The \href{https://g2.antv.vision/en/docs/api/general/scrollbar/}{official documentation}
for the list options to pass to \code{...}.
}
