% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspects.R
\name{select_asp}
\alias{select_asp}
\title{Select Specific Aspects}
\usage{
select_asp(asp, ...)
}
\arguments{
\item{asp}{Aspects as returned by \code{\link[=asp]{asp()}}.}

\item{...}{Names of aspects to select.}
}
\description{
Select specific aspects from a list of aspects
}
\examples{
\dontrun{
a <- asp(x = speed)

select_asp(a, "x")
}

}
\keyword{internal}
