% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{set_font}
\alias{set_font}
\title{Font}
\usage{
set_font(g2, font)
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}

\item{font}{Font family to use for chart text.}
}
\description{
Set chart font.
}
