% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slider.R
\name{slider}
\alias{slider}
\title{Slider}
\usage{
slider(g, ...)
}
\arguments{
\item{g}{An object of class \code{g2r} as returned by \code{\link[=g2]{g2()}}.}

\item{...}{Options to pass to the slider.}
}
\description{
Add a slider to the chart.
}
\seealso{
The \href{https://g2.antv.vision/en/docs/api/general/slider}{official documentation}
for the list options to pass to \code{...}.
}
