% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{smooth}
\alias{smooth}
\alias{fig_smooth}
\title{Smooth}
\usage{
fig_smooth(g2, ..., method = "lm", data = NULL, inherit_asp = TRUE,
  name = NULL)
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}

\item{...}{Any option, aesthetic (\code{\link{asp}}), or animation (\code{\link{Animation}}).}

\item{method}{Smoothing method (function) to use, accepts either a character vector, e.g. 
\code{"lm"}, \code{"glm"}, \code{"gam"}, \code{"loess"} or a function, e.g. \code{MASS::rlm}
 or \code{mgcv::gam}, \code{base::lm}, or \code{base::loess}.}

\item{data}{A \code{data.frame} containing data to chart or a \code{list}.}

\item{inherit_asp}{Whether to inherit aspects from \code{g2r}.}

\item{name}{Name of figure, useful to apply functions to specific figures.}
}
\description{
Add smoothing methods based on \code{x} and \code{y} and, optionally \code{group} aspects.
}
\examples{
g2(cars, asp(speed, dist)) \%>\% 
  fig_point() \%>\% 
  fig_smooth()

}
