% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{state}
\alias{state}
\alias{active}
\alias{selected}
\title{State}
\usage{
active(...)

selected(...)
}
\arguments{
\item{...}{Key value pair passed to styles.}
}
\description{
Customise the styles of figures given states
(active or selected).
}
\examples{
g2(iris, asp(Sepal.Width, Sepal.Length)) \%>\%
 fig_point(
   selected(fill = "red")
 ) \%>\% 
 conf_interplay("element", "selected")

}
