% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{states}
\alias{states}
\title{State}
\format{A \code{tibble} of 3 columns and 49 rows.
\describe{
  \item{\code{state}}{ US State}
  \item{\code{name}}{ A grouping variable}
  \item{\code{value}}{ Some count}
}}
\source{
Somewhere mysterious
}
\usage{
states
}
\description{
State and age data on something probably random.
}
\keyword{datasets}
