% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{subject}
\alias{subject}
\title{Subject}
\usage{
subject(g, subject, tag = htmltools::h3)
}
\arguments{
\item{g}{An object of class \code{g2r} or \code{g2Proxy} as returned by \code{\link[=g2]{g2()}} or
\code{\link[=g2_proxy]{g2_proxy()}}.}

\item{subject}{Subject of chart to display.}

\item{tag}{\code{htmltools} tag function to use.}
}
\description{
Add a subject to the plot.
}
\examples{
g2(cars, asp(speed, dist)) \%>\% 
 fig_point() \%>\% 
 subject("Points")

}
