% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{temp}
\alias{temp}
\title{Temperature}
\format{A \code{tibble} of 3 columns and 16 rows.
\describe{
  \item{\code{city}}{ City where temperature was taken}
  \item{\code{month}}{ Month}
  \item{\code{temp}}{ Temperature in Celsius}
}}
\source{
\url{https://antv.alipay.com/zh-cn/g2/3.x/demo/column/grouped-column.html}
}
\usage{
temp
}
\description{
Temperatures in London and Berlin, \emph{probably made up}.
}
\keyword{datasets}
