% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltip.R
\name{template}
\alias{template}
\alias{tpl}
\alias{tpl_item}
\title{Tooltip Template

Convenience function to create tooltip templates
(\code{itemTp} argument in \code{\link[=tooltip]{tooltip()}} function).}
\usage{
tpl(...)

tpl_item(name, value, marker = TRUE)
}
\arguments{
\item{...}{One or more \code{\link[=tpl_item]{tpl_item()}}.}

\item{name, value}{Name and value of the tooltip item.}

\item{marker}{Whether to include the color marker (dot)
in the tooltip.}
}
\description{
Tooltip Template

Convenience function to create tooltip templates
(\code{itemTp} argument in \code{\link[=tooltip]{tooltip()}} function).
}
\details{
The arguments \code{title}, \code{name}, and \code{value} accept
either a bare column name from the data to use as
\code{{mustache}}/\code{{handlebar}} in the template. If a string is
passed then it is treated as constant.
}
\examples{
template <- tpl(
 tpl_item(
   island,
   bill_depth_mm
 )
)

}
