% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme}
\alias{theme}
\alias{darken}
\alias{motif}
\title{Motif}
\usage{
darken(g2)

motif(g2, ...)
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}

\item{...}{Motif options.}
}
\description{
Customise the chart motif.
}
\examples{
g <- g2(iris, plan(Sepal.Length, Sepal.Width)) \%>\% 
  fig_point(plan(color = Species)) 

g \%>\% darken()

g \%>\% motif(colors = c("red", "green", "blue"))

}
