% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{to_g2r}
\alias{to_g2r}
\title{Convert to Tibble}
\usage{
to_g2r(data = NULL)
}
\arguments{
\item{data}{An object to convert.}
}
\description{
Converts objects to objects g2r can work with,
generally a \code{tibble::tibble}.
}
\details{
This is exposed so the user can understand
what happens under the hood and which variables/columns
can subsequently be used in figures with \code{\link[=asp]{asp()}}.

These methods are used in the \code{\link[=g2]{g2()}} function to preprocess
\code{data} objects.
}
\examples{
\dontrun{
to_g2r(AirPassengers)
}

}
