% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltip.R
\name{tooltip}
\alias{tooltip}
\title{Figure Tooltip}
\usage{
tooltip(callback, ...)
}
\arguments{
\item{callback}{A JavaScript callback function as
character string that accepts \code{asps} as input.}

\item{...}{Aspects to use for the \code{callback}.}
}
\description{
Control tooltip for a specific figure.
}
\details{
Define the aspects to use in the callback function.
This can be used in combination with \code{\link[=conf_tooltip]{conf_tooltip()}}.
}
\examples{
# callback function
cb <- "(mpg, qsec) => {
 return {
   name: mpg + ' (miles per gallon)',
   value: qsec + ' (1/4 mile time)'
 }
}"

g2(mtcars, asp(mpg, qsec)) \%>\% 
  fig_point(tooltip(cb, mpg, qsec))

# using a template
template <- "<p>{mpg} (mpg): {qsec} (1/4 mile time)</p>"

cb <- "(mpg, qsec) => {
 return {
   mpg,
   qsec
 }
}"

g2(mtcars, asp(mpg, qsec)) \%>\% 
 fig_point(tooltip(cb, mpg, qsec)) \%>\% 
 conf_tooltip(showTitle = FALSE, itemTpl = template)

}
