% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxies.R
\name{visible}
\alias{visible}
\alias{conceal}
\alias{reveal}
\alias{conceal_tooltip}
\alias{reveal_tooltip}
\title{Conceal & Reveal Figures}
\usage{
conceal(proxy, figures = NULL)

reveal(proxy, figures = NULL)

conceal_tooltip(proxy)

reveal_tooltip(proxy, x, y)
}
\arguments{
\item{proxy}{A g2r proxy as returned by \code{\link{g2Proxy}}}

\item{figures}{Index of figures to apply the change to.}

\item{x, y}{Coordinates of tooltip}
}
\description{
Conceal & reveal figures.

Conceal & reveal tooltips.
}
\examples{
library(shiny)

ui <- fluidPage(
  actionButton("hide", "hide"),
  actionButton("show", "show"),
  g2Output("chart")
)

server <- function(input, output) {
  output$chart <- renderG2({
    g2(cars, asp(speed, dist)) \%>\% 
      fig_point() \%>\% 
      fig_line()
  })

  observeEvent(input$hide, {
    g2Proxy("chart") \%>\% 
      conceal()
  })

  observeEvent(input$show, {
    g2Proxy("chart") \%>\% 
      reveal()
  })
}

\dontrun{shinyApp(ui, server)}

library(shiny)

ui <- fluidPage(
  actionButton("hide", "hide"),
  actionButton("show", "show"),
  g2Output("chart")
)

server <- function(input, output) {
  output$chart <- renderG2({
    g2(cars, asp(speed, dist)) \%>\% 
      fig_point() 
  })

  observeEvent(input$hide, {
    g2Proxy("chart") \%>\% 
      conceal_tooltip()
  })

  observeEvent(input$show, {
    g2Proxy("chart") \%>\% 
      reveal_tooltip(8, 16)
  })
}

\dontrun{shinyApp(ui, server)}

}
