% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.R
\name{adjust}
\alias{adjust}
\title{Adjust}
\usage{
adjust(type, margin = NULL, dodge_by = NULL)
}
\arguments{
\item{type}{A vector of types of adjustement to apply to the figure, see the "types" section below for valid values.}

\item{margin}{Margin, between \code{0} and \code{1}.}

\item{dodge_by}{Bare column name of}
}
\description{
Adjust a figure.
}
\section{Types}{

Valid values for the \code{type} argument.
\itemize{
  \item{\code{stack}}
  \item{\code{dodge}}
  \item{\code{jitter}}
  \item{\code{symmetric}}
}
}

