% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspects.R
\name{asp}
\alias{asp}
\title{Aspects}
\usage{
asp(x, y, ...)
}
\arguments{
\item{x, y}{Defines values to map to cartesian coordinates.}

\item{...}{Any other key value pair of aspect.}
}
\description{
Define aspects of a visualisation.
}
\section{Figure aspects}{

\itemize{
\item \code{x}, \code{y}
\item \code{ymin}, \code{ymax}
\item \code{size}
\item \code{color}
\item \code{shape}
\item \code{label}
\item \code{tooltip}
\item \code{style}
\item \code{interaction}
\item \code{color}
}
}

\section{Info aspects}{

\itemize{
\item \code{x}, \code{y}
\item \code{start}, \code{end}
\item \code{content}
\item \code{top}
}
}

