% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspects.R
\name{asp}
\alias{asp}
\title{Aspects}
\usage{
asp(x, y, ...)
}
\arguments{
\item{x, y, ...}{List of name value pairs giving aspects to map to
variables. The names for x and y aspects are typically omitted because
they are so common; all other aspects must be named.}
}
\description{
Chart's aspects, similar to \code{ggplot2}'s aesthetics.
}
\section{Aspects}{

\itemize{
  \item{\code{x}, \code{y}}
  \item{\code{size}}
  \item{\code{color}}
  \item{\code{shape}}
  \item{\code{opacity}}
  \item{\code{adjust}}
  \item{\code{tooltip}}
  \item{\code{label}}
  \item{\code{style}}
}
}

\examples{
g2(temp, asp(month, temp, color = city)) \%>\% 
  fig_interval(adjust("dodge"))

}
