% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{axis}
\alias{axis}
\alias{axis_x}
\alias{axis_y}
\alias{axis_asps}
\alias{axis_hide}
\title{Axes}
\usage{
axis_x(g, ...)

axis_y(g, ...)

axis_asps(g, asps, ...)

axis_hide(g)
}
\arguments{
\item{g}{An object of class \code{g2r} as returned by \code{\link[=g2]{g2()}}.}

\item{...}{Options to pass to the axis, passe \code{FALSE}
to hide the axis.}

\item{asps}{Column names to change the axis of.}
}
\description{
Customise the axes.
}
\section{Functions}{

\itemize{
\item \code{axis_x}: Customise the x axis.
\item \code{axis_y}: Customise the y axis.
\item \code{axis_asp}: Customise the axis by aspects (column names).
\item \code{axis_hide}: Hide all axis.
}
}

\examples{
g <- g2(cars, asp(speed, dist)) \%>\% 
 fig_point()

# hide axis
g \%>\% axis_x(FALSE)

# same as above
g \%>\% axis_asps("speed", FALSE)

# change position
g \%>\% axis_x(position = "top")

}
