% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{axis}
\alias{axis}
\alias{axis_x}
\alias{axis_y}
\alias{axis_cols}
\title{Axes}
\usage{
axis_x(g, ...)

axis_y(g, ...)

axis_cols(g, cols, ...)
}
\arguments{
\item{g}{An object of class \code{g2r} as returned by \code{\link[=g2]{g2()}}.}

\item{...}{Options to pass to the axis.}

\item{cols}{Column names to change the axis of.}
}
\description{
Customise the axes.
}
\section{Functions}{

\itemize{
\item \code{axis_x}: Customise the x axis.
\item \code{axis_y}: Customise the y axis.
\item \code{axis_cols}: Customise the axis by column names.
}
}

\examples{
g <- g2(cars, asp(speed, dist)) \%>\% 
 fig_point()

# hide axis
g \%>\% axis_x(FALSE)

# same as above
g \%>\% axis_cols("speed", FALSE)

# change position
g \%>\% axis_x(position = "top")

}
