% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R, R/legend.R
\name{axis}
\alias{axis}
\alias{axis_x}
\alias{axis_y}
\alias{axis_asps}
\alias{axis_hide}
\alias{legend_color}
\alias{legend_size}
\alias{legend_asps}
\title{Axes}
\usage{
axis_x(g, ...)

axis_y(g, ...)

axis_asps(g, asps, ...)

axis_hide(g)

legend_color(g, ...)

legend_size(g, ...)

legend_asps(g, asps, ...)
}
\arguments{
\item{g}{An object of class \code{g2r} or \code{g2Proxy} as returned by \code{\link[=g2]{g2()}} or
\code{\link[=g2_proxy]{g2_proxy()}}.}

\item{...}{Options to pass to the legend, pass \code{FALSE}
to hide the axis. Visit the
\href{https://g2.antv.vision/en/docs/api/general/legend}{official documentation}
for the full list of options.}

\item{asps}{Aspect (column names) to change the legend.}
}
\description{
Configure the axes.

Customise the legend.
}
\section{Functions}{

\itemize{
\item \code{axis_x}: Customise the x axis.
\item \code{axis_y}: Customise the y axis.
\item \code{axis_asps}: Customise the axis by aspects (column names).
\item \code{axis_hide}: Hide all axis.
}


\itemize{
\item \code{legend_color}: Customise the x axis.
\item \code{legend_size}: Customise the y axis.
\item \code{legend_asps}: Customise the axis by aspects (column names).
}
}

\examples{
g <- g2(cars, asp(speed, dist)) \%>\% 
 fig_point()

# hide axis
g \%>\% axis_x(FALSE)

# same as above
g \%>\% axis_asps("speed", FALSE)

# change position
g \%>\% axis_x(position = "top")

g <- g2(mtcars, asp(mpg, qsec, color = gear)) \%>\% 
 fig_point()

g \%>\% legend_color(position = "top")
g \%>\% legend_color(FALSE)

}
