% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxies.R
\name{change_data}
\alias{change_data}
\title{Change Data}
\usage{
change_data(proxy, data, ..., figures = NULL)
}
\arguments{
\item{proxy}{A g2r proxy as returned by \code{\link{g2Proxy}}}

\item{data}{A data.frame or \code{tibble}}

\item{...}{Bare names of column to select.}

\item{figures}{Index of figures to apply the change to.}
}
\description{
Change the data displayed on the chart.
}
\note{
The variable names must be identical to the initial data.
}
\examples{
library(shiny)

.make_data <- function(){
  dplyr::tibble(
    x = 1:20,
    y = runif(20)
  )
}

ui <- fluidPage(
  fluidRow(
    column(10, g2Output("chart")),
    column(2, actionButton("chg", "chg"))
  )
)

server <- function(input, output, session) {
  output$chart <- renderG2({
    g2(.make_data(), asp(x, y)) \%>\% 
      fig_point()
  })

  observeEvent(input$chg, {
    g2Proxy("chart") \%>\% 
      change_data(.make_data(), x, y)
  })
}

if(interactive()) shinyApp(ui, server)

}
