% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltip.R
\name{conf_tooltip}
\alias{conf_tooltip}
\title{Configure Tooltip}
\usage{
conf_tooltip(g, ...)
}
\arguments{
\item{g}{An object of class \code{g2r} as returned by \code{\link[=g2]{g2()}}.}

\item{...}{Options to pass to the axis, pass \code{FALSE}
to hide the axis. Visit the
\href{https://g2.antv.vision/en/docs/api/general/tooltip}{official documentation}
for the full list of options.}
}
\description{
Customise the tooltip.
}
\section{Functions}{

\itemize{
\item \code{axis_x}: Customise the x axis.
\item \code{axis_y}: Customise the y axis.
\item \code{axis_asps}: Customise the axis by aspects (column names).
\item \code{axis_hide}: Hide all axis.
}
}

\examples{
g2(mtcars, asp(drat, qsec, color = hp)) \%>\% 
 fig_point() \%>\% 
 conf_tooltip(
   showCrosshairs = TRUE,
   crosshairs = list(type = "xy")
)

}
