% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{conf_tooltip}
\alias{conf_tooltip}
\title{Tooltip}
\usage{
conf_tooltip(g2, ...)
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}

\item{...}{Configuration options or a logical indicating whether to show the tooltip.}
}
\description{
Configure the tooltip.
}
\examples{
g2(cars, asp(speed, dist, color = dist)) \%>\% 
  fig_point() \%>\% 
  conf_tooltip(crosshairs = "path")

}
