% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{coord}
\alias{coord}
\alias{conf_coord}
\alias{coord_rotate}
\alias{coord_scale}
\alias{coord_reflect}
\alias{coord_transpose}
\title{Coordinates}
\usage{
conf_coord(g2, type = c("rect", "polar", "theta", "helix"), ...)

coord_rotate(g2, angle = 90)

coord_scale(g2, sx, sy)

coord_reflect(g2, axis = "xy")

coord_transpose(g2)
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}

\item{type}{Type of coordinate axis.}

\item{...}{Any other options.}

\item{angle}{Angle of axis rotation.}

\item{sx, sy}{Scale of axis along \code{x} and \code{y} axis.}

\item{axis}{Axis to reflect (reverse).}
}
\description{
Configure chart coordinates axis.
}
\examples{
g2(cars, plan(speed, dist, color = dist)) \%>\% 
  fig_point() \%>\% 
  conf_coord("polar") 

}
