% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{coord}
\alias{coord}
\alias{coord_type}
\alias{coord_rotate}
\alias{coord_scale}
\alias{coord_reflect}
\alias{coord_transpose}
\title{Coordinates}
\usage{
coord_type(g2, type = c("rect", "polar", "theta", "helix"), ...)

coord_rotate(g2, angle = 90)

coord_scale(g2, sx, sy)

coord_reflect(g2, axis = "xy")

coord_transpose(g2)

coord(type = c("rect", "polar", "theta", "helix"), rotate = NULL,
  sx = NULL, sy = NULL, reflect = NULL, transpose = FALSE, ...)
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}

\item{type}{Type of coordinate axis.}

\item{...}{Any other options.}

\item{angle}{Angle of axis rotation.}

\item{sx, sy}{Scale of axis along \code{x} and \code{y} axis.}

\item{axis}{Axis to reflect (reverse).}

\item{coord, rotate, reflect, transpose}{Corresponds to arguments of other functions listed here.}
}
\description{
Configure chart coordinates axis.
}
\examples{
g2(cars, asp(speed, dist, color = dist)) \%>\% 
  fig_point() \%>\% 
  coord_type("helix")

g2(cars, asp(speed, dist, color = dist)) \%>\% 
  fig_point(coord("helix")) 

}
