% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{fig_bin}
\alias{fig_bin}
\title{Bins}
\usage{
fig_bin(g2, ..., type = c("rectangle", "hexagon"), bins = c(10, 5),
  size_count = TRUE, data = NULL, inherit_asp = TRUE, name = NULL)
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}

\item{...}{Any option, aesthetic (\code{\link{asp}}), or animation (\code{\link{Animation}}).}

\item{type}{Type of bins to create.}

\item{bins}{A vector or list defining the bin size.}

\item{size_count}{Whether to size the bins by count.}

\item{data}{A \code{data.frame} containing data to chart or a \code{list}.}

\item{inherit_asp}{Whether to inherit aspects from \code{g2r}.}

\item{name}{Name of figure, useful to apply functions to specific figures.}
}
\description{
Add bins based on \code{x}, and \code{y}.
}
\details{
Note that the function adds an aspect \code{count} which can be used, (see example).
}
\examples{
# issues warning in interactive session 
\dontrun{
g2(gaus, asp(x, y)) \%>\% 
  fig_bin(type = "hexagon")
}

}
