% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{fig_density}
\alias{fig_density}
\title{Density}
\usage{
fig_density(g2, ..., data = NULL, inherit_asp = TRUE, name = NULL)
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}

\item{...}{Any option, aesthetic (\code{\link{asp}}), or animation (\code{\link{Animation}}).}

\item{data}{A \code{data.frame} containing data to chart or a \code{list}.}

\item{inherit_asp}{Whether to inherit aspects from \code{g2r}.}

\item{name}{Name of figure, useful to apply functions to specific figures.}
}
\description{
Create a density plot based on \code{x} and optionally \code{group} aspects.
}
\examples{
g2(iris, asp(Petal.Length, color = Species)) \%>\% 
  fig_density()
 
}
