% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{fig_histogram}
\alias{fig_histogram}
\title{Histogram}
\usage{
fig_histogram(g2, ..., bin_width = 1, data = NULL,
  inherit_asp = TRUE, name = NULL)
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}

\item{...}{Any option, aesthetic (\code{\link{asp}}), or animation (\code{\link{Animation}}).}

\item{bin_width}{Width of bins.}

\item{data}{A \code{data.frame} containing data to chart or a \code{list}.}

\item{inherit_asp}{Whether to inherit aspects from \code{g2r}.}

\item{name}{Name of figure, useful to apply functions to specific figures.}
}
\description{
Create a histogram.
}
\examples{
g2(iris, asp(Sepal.Length, color = Species))  \%>\% 
  fig_histogram(bin_width = .3)

}
