% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{fig_map}
\alias{fig_map}
\title{Map}
\usage{
fig_map(g, ..., inherit_asp = TRUE, sync = TRUE, map = get_world_map())
}
\arguments{
\item{g}{An object of class \code{g2r} or \code{g2Proxy} as returned by \code{\link[=g2]{g2()}} or
\code{\link[=g2_proxy]{g2_proxy()}}.}

\item{...}{Options to pass to the figure, including \code{\link[=asp]{asp()}},
and \code{\link[=adjust]{adjust()}}, \code{\link[=active]{active()}}, \code{\link[=selected]{selected()}}, and \code{\link[=config]{config()}},
other key value pairs of are passed to
\href{https://g2.antv.vision/en/docs/api/general/style/}{style}.}

\item{inherit_asp}{Whether to inherit the aspects paseed to
\code{\link[=g2]{g2()}} initialisation function.}

\item{sync}{Whether to sync the axis data (align) with that
used in other figures, set to \code{FALSE} to not sync or set to
a character string to use as name of sync group.}

\item{map}{Name of map to pass to the \code{region} argument
of the \code{\link[=get_map_data]{get_map_data()}} function, or map object as returned by
\code{\link[=get_world_map]{get_world_map()}}, \code{\link[=get_gadm_data]{get_gadm_data()}}, or \code{\link[=get_map_data]{get_map_data()}},
or a \code{SpatialPolygonsDataFrame} as
returned by \code{\link[raster:getData]{raster::getData()}}, or a \code{geo_list} as
obtained from \code{geojsonio::geojson_list()}, or \code{MULTIPOLYGON}
of the class\code{sf} as obtained from reading shapefiles.}
}
\description{
Add a map figure.
}
\examples{
g2() \%>\%
  fig_map(stroke = "#fff", fill = "gray") \%>\%
  axis_hide()
}
