% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{fig_map}
\alias{fig_map}
\alias{get_map}
\title{Map}
\usage{
fig_map(g2, ..., map = NULL, axes = FALSE, tooltip = FALSE)

get_map(map = NULL)
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}

\item{...}{Additional options passed to \code{\link{fig_polygon}}.}

\item{map}{Name of country as defined by ISO Standard 3166-1, if \code{NULL} uses world map.}

\item{axes, tooltip}{Whether to show the tooltip and the axis.}
}
\description{
Add a map layer.
}
\examples{
\dontrun{
map <- get_map("fra")

g2(map, asp(longitude, latitude)) \%>\% 
  fig_polygon(axes = FALSE, tooltip = FALSE) \%>\% 
  style(height = cb("window.innerHeight"))

g2() \%>\% 
  fig_map(map = "fra") \%>\% 
  style(height = cb("window.innerHeight"))
}

}
