% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{fig_primitive}
\alias{fig_primitive}
\title{Path}
\usage{
fig_primitive(
  g,
  ...,
  data = NULL,
  inherit_asp = TRUE,
  sync = TRUE,
  type = c("point", "interval", "line", "schema", "edge", "area", "polygon")
)
}
\arguments{
\item{g}{An object of class \code{g2r} as returned by \code{\link[=g2]{g2()}}.}

\item{...}{Options to pass to the figure, including \code{\link[=asp]{asp()}},
and \code{\link[=adjust]{adjust()}}.}

\item{data}{A dataset (\code{data.frame} or \code{tibble}) to use to
draw the figure.}

\item{inherit_asp}{Whether to inherit the aspects paseed to
\code{\link[=g2]{g2()}} initialisation function.}

\item{sync}{Whether to sync the axis data (align).}
}
\description{
Add a path figure.
}
\keyword{internal}
