% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{fig_voronoi}
\alias{fig_voronoi}
\title{Voronoi}
\usage{
fig_voronoi(g2, ..., data = NULL, inherit_asp = TRUE, name = NULL)
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}

\item{...}{Any option, aesthetic (\code{\link{asp}}), or animation (\code{\link{Animation}}).}

\item{data}{A \code{data.frame} containing data to chart or a \code{list}.}

\item{inherit_asp}{Whether to inherit aspects from \code{g2r}.}

\item{name}{Name of figure, useful to apply functions to specific figures.}
}
\description{
Create a voronoi plot.
}
\examples{
df <- dplyr::tibble(
  x = runif(25, 1, 500),
  y = runif(25, 1, 500),
  value = runif(25, 1, 500)
)

g2(df, asp(x, y, color = value)) \%>\% 
  fig_voronoi()

}
