% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{fig_waffle}
\alias{fig_waffle}
\title{Waffle}
\usage{
fig_waffle(g2, ..., rows = 10, count = 500, data = NULL,
  inherit_asp = TRUE, name = NULL)
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}

\item{...}{Any option, aesthetic (\code{\link{asp}}), or animation (\code{\link{Animation}}).}

\item{rows}{Number of rows.}

\item{count}{Number of squares to use.}

\item{data}{A \code{data.frame} containing data to chart or a \code{list}.}

\item{inherit_asp}{Whether to inherit aspects from \code{g2r}.}

\item{name}{Name of figure, useful to apply functions to specific figures.}
}
\description{
Create a waffle plot based on \code{x}, \code{y}, and optionally \code{color} aspects.
}
\examples{
fruits \%>\% 
  dplyr::mutate(value = value * 100) \%>\% 
  g2(asp(fruit, value)) \%>\% 
 fig_waffle(asp(color = fruit))

}
