% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{figures}
\alias{figures}
\alias{fig_interval}
\title{Interval}
\usage{
fig_interval(
  g,
  ...,
  sync = TRUE,
  data = NULL,
  inherit_asp = TRUE,
  style = NULL
)
}
\arguments{
\item{g}{An object of class \code{g2r} or \code{g2Proxy} as returned by \code{\link[=g2]{g2()}} or
\code{\link[=g2_proxy]{g2_proxy()}}.}

\item{...}{Options to pass to the figure, including \code{\link[=asp]{asp()}},
and \code{\link[=adjust]{adjust()}}, \code{\link[=active]{active()}}, and \code{\link[=selected]{selected()}}.}

\item{sync}{Whether to sync the axis data (align) with that
used in other figures, set to \code{FALSE} to not sync or set to
a character string to use as name of sync group.}

\item{data}{A dataset (\code{data.frame} or \code{tibble}) to use to
draw the figure.}

\item{inherit_asp}{Whether to inherit the aspects paseed to
\code{\link[=g2]{g2()}} initialisation function.}

\item{style}{A list of options to pass to
\href{https://g2.antv.vision/en/docs/api/general/style}{style} that
customise the figure.}
}
\description{
Add an interval figure.
}
\examples{
g2(sleep, asp(ID, extra, color = group)) \%>\% 
 fig_interval()

df <- data.frame(
 cat = letters[1:5],
 value = c(0.15, .3, .65, .75, .9)
)

g2(df, asp(cat, value, color = cat, shape = "funnel")) \%>\% 
 fig_interval(adjust("symmetric")) \%>\% 
 coord_type("rect") \%>\% 
 coord_transpose() \%>\% 
 coord_scale(-1, 1) \%>\% 
 axis_hide()

}
