% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2r.R
\name{g2}
\alias{g2}
\title{Initialise}
\usage{
g2(data = NULL, ..., width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{A data.frame or tibble containing data to chart.}

\item{...}{Aspects of the chart, see \code{\link[=asp]{asp()}}.}

\item{width, height}{Dimensions of the chart, accepts
any valid CSS unit e.g.: \verb{100\%}, numerics are treated
as pixels, e.g.: \code{400} = \verb{400px}.}

\item{elementId}{Valid CSS id attribute.}
}
\description{
Initialise a chart.
}
\examples{
g2(cars) \%>\% 
 fig_point(asp(speed, dist))

}
