% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{g2_action}
\alias{g2_action}
\title{Plot Action}
\usage{
g2_action(plot_id, btn_id, ..., data = NULL)
}
\arguments{
\item{plot_id}{Id of chart to interact with.}

\item{btn_id}{Id of the \code{\link[=input_button]{input_button()}} that triggers the action.}

\item{...}{Aspects, see \code{\link[=asp]{asp()}}.}

\item{data}{Data.frame containing data to plot.}
}
\description{
Include dynamic elements in Rmarkdown.
}
