% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2.R
\name{g2_chart}
\alias{g2_chart}
\title{G2 Chart}
\usage{
g2_chart(data = NULL, padding = NULL, forceFit = TRUE, animate = TRUE,
  options = list(), width = NULL, height = NULL, elementId = NULL)
}
\description{
Create a 'g2' chart
}
