% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{g2_geom}
\alias{g2_geom}
\title{<Add Title>}
\usage{
g2_geom(g2, x, y, type, color = NULL, adjust = NULL, shape = NULL,
  size = NULL, opacity = NULL, tooltip = NULL, label = NULL,
  style = NULL, select = NULL, active = NULL)
}
\description{
<Add Description>
}
