% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2r.R
\name{g2_proxy}
\alias{g2_proxy}
\title{Shiny Proxy}
\usage{
g2_proxy(id, ..., data = NULL, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{Id of chart to interact with.}

\item{...}{Aspects, see \code{\link[=asp]{asp()}}.}

\item{data}{Data.frame containing data to plot.}

\item{session}{A valid shiny session.}
}
\description{
Proxy to dynamically interact with the chart in shiny.
}
