% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{g2_scale_shape}
\alias{g2_scale_shape}
\title{Scale shape}
\usage{
g2_scale_shape(g2, shapes)
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}

\item{shapes}{A vector of shapes or a JavaScript function.}
}
\description{
Scale shapes.
}
