% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2r.R
\name{g2r}
\alias{g2r}
\title{Initialise}
\usage{
g2r(data = NULL, mapping = NULL, ..., render = TRUE, width = NULL,
  height = NULL, elementId = NULL)
}
\arguments{
\item{data}{A \code{data.frame} containing data to chart.}

\item{mapping}{Mapping aesthetics as returned by \code{\link{gaes}}.}

\item{...}{Any general options.}

\item{render}{Whether to render the chart.}

\item{width, height}{Dimensions of chart.}

\item{elementId}{ID of \code{DOM} container.}
}
\description{
Initialise a g2 chart.
}
\examples{
cars \%>\% 
  g2r(gaes(speed, dist)) \%>\% 
  g2_point()

}
