% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauge.R
\name{gauge}
\alias{gauge}
\alias{gauge_x}
\alias{gauge_y}
\alias{gauge_color}
\alias{gauge_colour}
\alias{gauge_size}
\alias{gauge_cols}
\title{Gauge}
\usage{
gauge_x(g, ...)

gauge_y(g, ...)

gauge_color(g, ...)

gauge_colour(g, ...)

gauge_size(g, ...)

gauge_cols(g, cols, ...)
}
\description{
Gauge
}
