% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{gauge}
\alias{gauge}
\alias{gauge_x_time}
\alias{gauge_y_time}
\alias{gauge_x_linear}
\alias{gauge_y_linear}
\alias{gauge_x_cat}
\alias{gauge_y_cat}
\alias{gauge_x_time_cat}
\alias{gauge_y_time_cat}
\alias{gauge_x_log}
\alias{gauge_y_log}
\alias{gauge_x_pow}
\alias{gauge_y_pow}
\alias{gauge_x_quantile}
\alias{gauge_y_quantile}
\alias{gauge_x_quantize}
\alias{gauge_y_quantize}
\alias{gauge_x_identity}
\alias{gauge_y_identity}
\alias{gauge_asp}
\title{Gauge Grid}
\usage{
gauge(
  g,
  asp,
  ...,
  nice = TRUE,
  range = NULL,
  min = NULL,
  max = NULL,
  min_limit = NULL,
  max_limit = NULL,
  alias = NULL,
  tick_count = NULL,
  max_tick_count = NULL
)

gauge_x_time(g, ..., show_last = FALSE)

gauge_y_time(g, ..., show_last = FALSE)

gauge_x_linear(g, ..., tick_interval = NULL)

gauge_y_linear(g, ..., tick_interval = NULL)

gauge_x_cat(g, ...)

gauge_y_cat(g, ...)

gauge_x_time_cat(g, ...)

gauge_y_time_cat(g, ...)

gauge_x_log(g, ..., base = 10)

gauge_y_log(g, ..., base = 10)

gauge_x_pow(g, ...)

gauge_y_pow(g, ...)

gauge_x_quantile(g, ...)

gauge_y_quantile(g, ...)

gauge_x_quantize(g, ...)

gauge_y_quantize(g, ...)

gauge_x_identity(g, ...)

gauge_y_identity(g, ...)

gauge_asp(g, ...)
}
\arguments{
\item{g}{An object of class \code{g2r} or \code{g2Proxy} as returned by \code{\link[=g2]{g2()}} or
\code{\link[=g2_proxy]{g2_proxy()}}.}

\item{asp}{Bare column name of aspect to apply the
gauge to.}

\item{...}{Options to gauge variables
(\code{aspects} defined by \code{\link[=asp]{asp()}}).}

\item{nice}{Automatically adjust \code{min}, and \code{max}.}

\item{range}{A vector of length 2 giving the minimum, and
maximum.}

\item{min, max}{Range of the gauge.}

\item{min_limit, max_limit}{Strict range of the ticks.}

\item{alias}{Alias name of the gauge and variable to
display.}

\item{tick_count}{Maximum number of ticks.}

\item{max_tick_count}{Maximum number of ticks.}

\item{show_last}{Whether to force show the last tick
(only for \code{time} gauges).}

\item{tick_interval}{Minimum tick interval, only applies to
linear type of gauge.}

\item{base}{Base of log.}
}
\description{
Gauge the variables (\code{aspects}) used to define axis
and grid of the plot.
}
\section{Types}{

\itemize{
\item \code{cat}: Categorical.
\item \code{timeCat}: Categorical time.
\item \code{linear}: Linear.
\item \code{time}: Date, time, etc.
\item \code{log}: Logarithmic.
\item \code{pow}: Exponential.
\item \code{quantize}: Manual quantiles.
\item \code{quantile}: Auto-generated quantiles.
\item \code{identity}: Constant.
}
}

\examples{
g <- g2(cars, asp(speed, dist)) \%>\%
  fig_point()

g \%>\% gauge(speed, min = 0)
g \%>\% gauge_y_log(title = "Log")
g \%>\% gauge(dist, tickCount = 10)
}
