% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauge.R
\name{gauge_}
\alias{gauge_}
\title{Upsert Gauge}
\usage{
gauge_(g, cols, ...)
}
\arguments{
\item{g}{An object of class \code{g2r} as returned by \code{\link[=g2]{g2()}}.}

\item{cols}{Columns to apply the options to.}

\item{...}{Options to pass to the gauge.}
}
\description{
Upsert gauge for given column.
}
\keyword{internal}
