% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauge.R
\name{gauge_color}
\alias{gauge_color}
\title{Gauge colour}
\usage{
gauge_color(g2, colors = NULL, callback = NULL)
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}

\item{colors}{A vector of colors.}

\item{callback}{A JavaScript callback function (see \code{\link{cb}}) which returns a color.}
}
\description{
Gauge colour, similar to the \code{scale_colour_*} family of functions from the \code{ggplot2} package.
}
