% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauge.R
\name{gauge_opacity}
\alias{gauge_opacity}
\title{Gauge opacity}
\usage{
gauge_opacity(g2, callback = NULL)
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}

\item{callback}{A JavaScript callback function (see \code{\link{cb}}) which returns a color.}
}
\description{
Gauge opacity.
}
\examples{
func <- cb("function(dist){
  if(dist > 60)
    return 1
  return .3
}")

g2(cars, asp(speed, dist, opacity = dist)) \%>\% 
  fig_point() \%>\% 
  gauge_opacity(callback = func)

}
