% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauge.R
\name{gauge_size}
\alias{gauge_size}
\title{Gauge size}
\usage{
gauge_size(g2, size)
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}

\item{size}{A vector of sizes (\code{min, max}) or a JavaScript function.}
}
\description{
Gauge size.
}
