% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauge.R
\name{gauge_tooltip}
\alias{gauge_tooltip}
\title{Gauge tooltip}
\usage{
gauge_tooltip(g2, callback = NULL)
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}

\item{callback}{A JavaScript callback function (see \code{\link{cb}}) which returns a color.}
}
\description{
Gauge tooltip.
}
