% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauge.R
\name{gauges_types}
\alias{gauges_types}
\title{Set Gauges Types}
\usage{
gauges_types(g, cols, data = NULL)
}
\arguments{
\item{g}{An object of class \code{g2r} as returned by \code{\link[=g2]{g2()}}.}

\item{cols}{Column names to set.}

\item{data}{Dataset to use to determine types.}
}
\description{
Set Gauges Types
}
\keyword{internal}
