% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauge.R
\name{gaugeViews}
\alias{gaugeViews}
\alias{gauge_color}
\alias{gauge_size}
\alias{gauge_shape}
\alias{gauge_label}
\alias{gauge_tooltip}
\alias{gauge_style}
\title{Gauge Aspects}
\usage{
gauge_color(g, handler)

gauge_size(g, handler)

gauge_shape(g, handler)

gauge_label(g, handler)

gauge_tooltip(g, handler)

gauge_label(g, handler)

gauge_style(g, handler)
}
\arguments{
\item{g}{An object of class \code{g2r} as returned by \code{\link[=g2]{g2()}}.}

\item{handler}{Handler to customise the aspect.
Generally a vector or a function (wrapped in \code{htmlwidgets::JS}).}
}
\description{
Customise aspects of the chart.
}
\examples{
# base plot
g <- g2(cars, asp(speed, dist)) \%>\% 
 fig_point(asp(color = speed)) 

# color with vector
g \%>\% gauge_color(c("red", "white", "blue"))

# color with callback
cb <- "function(mpg){
 if(mpg > 25){
   return 'blue';
 }

 return 'red';
}"

g \%>\% gauge_color(htmlwidgets::JS(cb))

}
\seealso{
\link{gauge} to gauge aspects of the grid and axis.
}
