% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{geoms}
\alias{geoms}
\alias{fig_interval}
\alias{fig_line}
\alias{fig_point}
\alias{fig_path}
\alias{fig_area}
\alias{fig_polygon}
\alias{fig_schema}
\alias{fig_edge}
\alias{fig_violin}
\alias{fig_heatmap}
\alias{fig_point_jitter}
\alias{fig_point_stack}
\alias{fig_point_dodge}
\alias{fig_interval_stack}
\alias{fig_interval_dodge}
\alias{fig_interval_symmetric}
\alias{fig_area_stack}
\alias{fig_schema_dodge}
\alias{fig_venn}
\title{Primitive Firgures}
\usage{
fig_interval(g2, ..., data = NULL, inherit_asp = TRUE, name = NULL)

fig_line(g2, ..., data = NULL, inherit_asp = TRUE, name = NULL)

fig_point(g2, ..., data = NULL, inherit_asp = TRUE, name = NULL)

fig_path(g2, ..., data = NULL, inherit_asp = TRUE, name = NULL)

fig_area(g2, ..., data = NULL, inherit_asp = TRUE, name = NULL)

fig_polygon(g2, ..., data = NULL, inherit_asp = TRUE, name = NULL)

fig_schema(g2, ..., data = NULL, inherit_asp = TRUE, name = NULL)

fig_edge(g2, ..., data = NULL, inherit_asp = TRUE, name = NULL)

fig_violin(g2, ..., data = NULL, inherit_asp = TRUE, name = NULL)

fig_heatmap(g2, ..., data = NULL, inherit_asp = TRUE, name = NULL)

fig_point_jitter(g2, ..., data = NULL, inherit_asp = TRUE,
  name = NULL)

fig_point_stack(g2, ..., data = NULL, inherit_asp = TRUE,
  name = NULL)

fig_point_dodge(g2, ..., data = NULL, inherit_asp = TRUE,
  name = NULL)

fig_interval_stack(g2, ..., data = NULL, inherit_asp = TRUE,
  name = NULL)

fig_interval_dodge(g2, ..., data = NULL, inherit_asp = TRUE,
  name = NULL)

fig_interval_symmetric(g2, ..., data = NULL, inherit_asp = TRUE,
  name = NULL)

fig_area_stack(g2, ..., data = NULL, inherit_asp = TRUE, name = NULL)

fig_schema_dodge(g2, ..., data = NULL, inherit_asp = TRUE,
  name = NULL)

fig_venn(g2, ..., data = NULL, inherit_asp = TRUE, name = NULL)
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}

\item{...}{Any option, aesthetic (\code{\link{asp}}), or animation (\code{\link{Animation}}).}

\item{data}{A \code{data.frame} containing data to chart or a \code{list}.}

\item{inherit_asp}{Whether to inherit aspects from \code{g2r}.}

\item{name}{Name of figure, useful to apply functions to specific figures.}
}
\description{
Figures, equivalent to \code{ggplot2}'s geometries.
}
\examples{
g2(cars, asp(dist, speed)) \%>\% 
  fig_point() \%>\% 
  fig_line()

g2(fruits, asp(value, color = fruit)) \%>\% 
  fig_interval_stack() \%>\% 
  coord_type("theta")

}
