% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geoms}
\alias{geoms}
\alias{g2_interval}
\alias{g2_line}
\alias{g2_point}
\alias{g2_path}
\alias{g2_area}
\alias{g2_polygon}
\alias{g2_schema}
\alias{g2_edge}
\alias{g2_heatmap}
\alias{g2_point_jitter}
\alias{g2_point_stack}
\alias{g2_point_dodge}
\alias{g2_interval_stack}
\alias{g2_interval_dodge}
\alias{g2_interval_symmetric}
\alias{g2_area_stack}
\alias{g2_schema_dodge}
\title{Bar}
\usage{
g2_interval(g2, ..., data = NULL, inherit_gaes = TRUE)

g2_line(g2, ..., data = NULL, inherit_gaes = TRUE)

g2_point(g2, ..., data = NULL, inherit_gaes = TRUE)

g2_path(g2, ..., data = NULL, inherit_gaes = TRUE)

g2_area(g2, ..., data = NULL, inherit_gaes = TRUE)

g2_polygon(g2, ..., data = NULL, inherit_gaes = TRUE)

g2_schema(g2, ..., data = NULL, inherit_gaes = TRUE)

g2_edge(g2, ..., data = NULL, inherit_gaes = TRUE)

g2_heatmap(g2, ..., data = NULL, inherit_gaes = TRUE)

g2_point_jitter(g2, ..., data = NULL, inherit_gaes = TRUE)

g2_point_stack(g2, ..., data = NULL, inherit_gaes = TRUE)

g2_point_dodge(g2, ..., data = NULL, inherit_gaes = TRUE)

g2_interval_stack(g2, ..., data = NULL, inherit_gaes = TRUE)

g2_interval_dodge(g2, ..., data = NULL, inherit_gaes = TRUE)

g2_interval_symmetric(g2, ..., data = NULL, inherit_gaes = TRUE)

g2_area_stack(g2, ..., data = NULL, inherit_gaes = TRUE)

g2_schema_dodge(g2, ..., data = NULL, inherit_gaes = TRUE)
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}

\item{...}{Any option or aesthetic.}

\item{data}{A \code{data.frame} containing data to chart.}

\item{inherit_gaes}{Whether to inherit aesthetics from \code{g2r}.}
}
\description{
Bar
}
\examples{
cars \%>\% 
  g2r(gaes(speed, dist)) \%>\% 
  g2_point()

}
