% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{info}
\alias{info}
\alias{info_text}
\alias{info_image}
\alias{info_arc}
\alias{info_line}
\alias{info_region}
\alias{info_region_filter}
\alias{info_marker}
\alias{info_data_region}
\alias{info_shape}
\alias{info_html}
\title{Text}
\usage{
info_text(g, ..., style = NULL)

info_image(g, ..., style = NULL)

info_arc(g, ..., style = NULL)

info_line(g, ..., style = NULL)

info_region(g, ..., style = NULL)

info_region_filter(g, ..., style = NULL)

info_marker(g, ..., style = NULL)

info_data_region(g, ..., style = NULL)

info_shape(g, ..., style = NULL)

info_html(g, ..., style = NULL)
}
\arguments{
\item{g}{An object of class \code{g2r} as returned by \code{\link[=g2]{g2()}}.}

\item{...}{Options to pass to the informational annotation.}

\item{style}{A list of options defning the style.}
}
\description{
Add a point figure.
}
\examples{
g2(cars, asp(speed, dist)) \%>\%
 fig_point() \%>\% 
 info_text(
   position = c(5, 4), content = "Look here!"
 )

}
\seealso{
\href{https://antv-g2.gitee.io/en/docs/api/general/annotation}{Official annotation documentation}
for defails pon what to pass to \code{...}.
}
