% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{info}
\alias{info}
\alias{info_line}
\alias{info_vline}
\alias{info_hline}
\alias{info_text}
\alias{info_image}
\alias{info_region}
\alias{info_html}
\alias{info_arc}
\alias{info_data_marker}
\alias{info_region_filter}
\alias{info_data_region}
\title{Info}
\usage{
info_line(g2, ..., data = NULL, figures = NULL, inherit_asp = TRUE)

info_vline(g2, ..., data = NULL, figures = NULL, inherit_asp = TRUE)

info_hline(g2, ..., data = NULL, figures = NULL, inherit_asp = TRUE)

info_text(g2, ..., data = NULL, figures = NULL, inherit_asp = TRUE)

info_image(g2, ..., data = NULL, figures = NULL, inherit_asp = TRUE)

info_region(g2, ..., data = NULL, figures = NULL, inherit_asp = TRUE)

info_html(g2, ..., data = NULL, figures = NULL, inherit_asp = TRUE)

info_arc(g2, ..., data = NULL, figures = NULL, inherit_asp = TRUE)

info_data_marker(g2, ..., data = NULL, figures = NULL,
  inherit_asp = TRUE)

info_region_filter(g2, ..., data = NULL, figures = NULL,
  inherit_asp = TRUE)

info_data_region(g2, ..., data = NULL, figures = NULL,
  inherit_asp = TRUE)
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}

\item{...}{Info options and \code{\link{asp}}.}

\item{data}{A \code{data.frame} containing data for \code{\link{asp}} if \code{NULL} is inherited from \code{\link{g2}}.}

\item{figures}{Vector of names or indices of figure(s) to apply the information to, if \code{NULL} applies only to the first figure.}

\item{inherit_asp}{Whether to inherit aspects from \code{g2r}.}
}
\description{
Add informational elements to the chart.
}
\examples{
g2(mtcars, asp(mpg, qsec)) \%>\% 
  fig_point() \%>\% 
  info_data_marker(
    content = "Marker",
    position = c(20, 20)
  )

newcars <- cars[1:26,]
newcars$text <- LETTERS

g2(cars, asp(speed, dist)) \%>\% 
  fig_point() \%>\% 
  info_text(asp(speed, dist, content = text), data = newcars, offsetY = -20)
 
}
