% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{interplay}
\alias{interplay}
\title{Interplay}
\usage{
interplay(g, ...)
}
\arguments{
\item{g}{An object of class \code{g2r} or \code{g2Proxy} as returned by \code{\link[=g2]{g2()}} or
\code{\link[=g2_proxy]{g2_proxy()}}.}

\item{...}{String(s) defining interactions.}
}
\description{
Add interplay (interactions) to the chart.
}
\examples{
df <- data.frame(
 x = letters,
 y = runif(26)
)

g2(df, asp(x, y)) \%>\%
 fig_interval(
   selected(fill = "orange")
 ) \%>\% 
 interplay("element", "selected")

}
