% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{motif}
\alias{motif}
\title{Motif}
\usage{
motif(
  g,
  ...,
  brandColor = NULL,
  backgroundColor = "transparent",
  renderer = c("canvas", "svg"),
  padding = "auto",
  visible = TRUE
)
}
\arguments{
\item{g}{An object of class \code{g2r} or \code{g2Proxy} as returned by \code{\link[=g2]{g2()}} or
\code{\link[=g2_proxy]{g2_proxy()}}.}

\item{...}{Key value pair defining style, or
\code{\link[=element]{element()}}.}

\item{brandColor}{Main default color.}

\item{backgroundColor}{Plot background color.

The styling options are poorly, if at all
documented online.}

\item{renderer}{Renderer to use, defaults to \code{canvas}.}

\item{padding}{An integer, or a vector of length 4.}

\item{visible}{Whether the chart is visible.}
}
\description{
Set the motif of the chart, defaults to \code{light}.
}
\examples{
g2(iris, asp(Sepal.Width, Sepal.Length)) \%>\% 
 fig_point(
   asp(color = Species, shape = "circle")
 ) \%>\% 
 motif(
   brandColor = "orange",
   backgroundColor = "black",
   elementPoint(
     shape = "circle",
     stroke = "white",
     fiillOpacity = .7
   )
 )

}
