% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{motif}
\alias{motif}
\alias{motif_global}
\title{Motif}
\usage{
motif(
  g,
  ...,
  renderer = c("canvas", "svg"),
  padding = "auto",
  visible = TRUE,
  name = "light"
)

motif_global(
  ...,
  renderer = c("canvas", "svg"),
  padding = "auto",
  visible = TRUE,
  name = "light"
)
}
\arguments{
\item{g}{An object of class \code{g2r} or \code{g2Proxy} as returned by \code{\link[=g2]{g2()}} or
\code{\link[=g2_proxy]{g2_proxy()}}.}

\item{...}{Key value pair defining style, please see
some the options available on the
\href{https://g2.antv.vision/en/docs/api/advanced/register-theme}{official documentation}.}

\item{renderer}{Renderer to use, defaults to \code{canvas}.}

\item{padding}{An integer, or a vector of length 4.}

\item{visible}{Whether the chart is visible.}

\item{name}{Name of the motif to apply to charts.
Out-of-the-box g2 provides \code{light}, and \code{dark}.}
}
\description{
Set the motif of the chart, defaults to \code{light}.
}
\details{
\code{\link[=motif]{motif()}} applies the theme to the chart, \code{\link[=motif_global]{motif_global()}}
creates a global theme that all subsequent charts will use.

The styling options are poorly, if at all
documented online.
}
\examples{
# mimic ggplot2
g2(iris, asp(Sepal.Width, Sepal.Length)) \%>\% 
 fig_point(
   asp(color = Species, shape = "circle")
 ) \%>\% 
 motif(
   background = "darkgrey",
   colors10 = c(
     "#F8766D", 
     "#7CAE00", 
     "#00BFC4", 
     "#C77CFF"
   )
 )

}
