% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{motif}
\alias{motif}
\alias{global_motif}
\alias{motif_from_json}
\alias{motif_from_list}
\alias{motif_as_list}
\title{Motif}
\usage{
motif(
  g,
  ...,
  brandColor = NULL,
  backgroundColor = "transparent",
  renderer = c("canvas", "svg"),
  padding = "auto",
  visible = TRUE
)

global_motif(
  ...,
  brandColor = NULL,
  backgroundColor = "transparent",
  renderer = c("canvas", "svg"),
  padding = "auto",
  visible = TRUE
)

motif_from_json(g, path)

motif_from_list(g, motif)

motif_as_list(..., brandColor = NULL, backgroundColor = "transparent")
}
\arguments{
\item{g}{An object of class \code{g2r} or \code{g2Proxy} as returned by \code{\link[=g2]{g2()}} or
\code{\link[=g2_proxy]{g2_proxy()}}.}

\item{...}{Key value pair defining style, or
\code{\link[=element]{element()}}.}

\item{brandColor}{Main default color.}

\item{backgroundColor}{Plot background color.}

\item{renderer}{Renderer to use, defaults to \code{canvas}.}

\item{padding}{An integer, or a vector of length 4.}

\item{visible}{Whether the chart is visible.}

\item{path}{Path to JSON file.}

\item{motif}{List defnining the theme, similar to JSON.}
}
\description{
Set the motif of the chart, defaults to \code{light}.
}
\details{
The function \code{\link[=motif_from_json]{motif_from_json()}} can be used to
define the theme from a JSON file of theme, to see the
default theme file:
\code{system.file("theme.json", package = "g2r")}.
}
\section{Functions}{

\itemize{
\item \code{motif}: Defines the motif of a visualisation.
\item \code{motif_from_json}: Defines the motif from a JSON
file of theme, see the theme file.
\code{system.file("theme.json", package = "g2r")}
\item \code{motif_from_list}: Defines the motif from a \code{list},
derived from the JSON file.
\item \code{motif_as_list}: Returns a motif as a \code{list} to use
with \code{\link[=motif_from_list]{motif_from_list()}}.
\item \code{global_motif}: Define a global motif that will be used by
all subsequent charts.
}
}

\examples{
g2(iris, asp(Sepal.Width, Sepal.Length)) \%>\%
  fig_point(
    asp(color = Species, shape = "circle")
  ) \%>\%
  motif(
    brandColor = "orange",
    backgroundColor = "black",
    elementPoint(
      shape = "circle",
      stroke = "white",
      fillOpacity = .7
    )
  )
}
