% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{motif}
\alias{motif}
\title{Style}
\usage{
motif(g2, coord_type = NULL, coord_rotate = NULL, coord_sx = NULL,
  coord_sy = NULL, coord_reflect = NULL, coord_transpose = NULL,
  width = NULL, height = NULL, padding = rep(10, 4),
  bg_fill = NULL, bg_opacity = NULL, bg_fill_opacity = NULL,
  bg_stroke = NULL, bg_stroke_opacity = NULL, bg_line_width = NULL,
  bg_radius = NULL, plot_fill = NULL, plot_fill_opacity = NULL,
  plot_stroke = NULL, plot_stroke_opacity = NULL,
  plot_opacity = NULL, plot_line_width = NULL, plot_radius = NULL,
  fit = TRUE, animate = TRUE, pixel_ratio = NULL, renderer = NULL,
  font = NULL)
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}

\item{coord_type, coord_rotate, coord_sx, coord_sy, coord_reflect, coord_transpose}{Customise global coordinates of chart.}

\item{width, height}{Chart dimensions.}

\item{padding}{Padding around chart, a vector of length \code{4} corresponding to top, right, bottom, and left.}

\item{bg_fill, bg_opacity, bg_fill_opacity, bg_stroke, bg_stroke_opacity, bg_line_width, bg_radius}{Customise background of global chart.}

\item{plot_fill, plot_fill_opacity, plot_stroke, plot_stroke_opacity, plot_opacity, plot_line_width, plot_radius}{Customise plot background.}

\item{fit}{Whether to force graph to fit parent element.}

\item{animate}{Whether to animate the chart.}

\item{pixel_ratio}{Device pixel ratio, defaults to \code{window.devicePixelRatio}.}

\item{renderer}{Renderer, \code{canvas} or \code{svg}.}

\item{font}{Font to use on chart.}
}
\description{
Customise general style of the chart.
}
\examples{
g2(cars, asp(speed, dist)) \%>\% 
  fig_point() \%>\% 
  motif(plot_fill = "grey")

}
