% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{order_data}
\alias{order_data}
\title{Order Data}
\usage{
order_data(data = NULL, x = NULL, color = NULL)
}
\arguments{
\item{data}{Data to reorder.}

\item{x}{Column on x axis.}

\item{color}{Color variable name.}
}
\description{
Fortunately or unfortunately G2.js does not reorder
the data and uses it as-is.
}
\details{
This function reorders along the \code{x} axis to make sure
most plots look correct, see the \code{NO_REORDER_TYPES} for
figures types that are excluded from reordering.
}
\keyword{internal}
