% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauge.R
\name{palettes}
\alias{palettes}
\alias{gauge_color_viridis}
\alias{gauge_color_plasma}
\alias{gauge_color_inferno}
\alias{gauge_color_magma}
\alias{gauge_color_accent}
\alias{gauge_color_dark2}
\alias{gauge_color_paired}
\alias{gauge_color_pastel1}
\alias{gauge_color_pastel2}
\alias{gauge_color_set1}
\alias{gauge_color_set2}
\alias{gauge_color_set3}
\alias{gauge_color_neon}
\alias{gauge_color_std}
\alias{gauge_color_pink}
\alias{gauge_color_orange}
\alias{gauge_color_lime}
\alias{gauge_color_blue}
\alias{gauge_color_aw}
\alias{gauge_color_g2rq}
\alias{gauge_color_g2rc}
\alias{gauge_color_g2rd}
\alias{gauge_color_brbg}
\alias{gauge_color_piyg}
\alias{gauge_color_prgn}
\alias{gauge_color_puor}
\alias{gauge_color_rdbu}
\alias{gauge_color_rdgy}
\alias{gauge_color_rdylbu}
\alias{gauge_color_rdylgn}
\alias{gauge_color_spectral}
\title{Color Palettes}
\usage{
gauge_color_viridis(g)

gauge_color_plasma(g)

gauge_color_inferno(g)

gauge_color_magma(g)

gauge_color_accent(g)

gauge_color_dark2(g)

gauge_color_paired(g)

gauge_color_pastel1(g)

gauge_color_pastel2(g)

gauge_color_set1(g)

gauge_color_set2(g)

gauge_color_set3(g)

gauge_color_neon(g)

gauge_color_std(g)

gauge_color_pink(g)

gauge_color_orange(g)

gauge_color_lime(g)

gauge_color_blue(g)

gauge_color_aw(g)

gauge_color_g2rq(g)

gauge_color_g2rc(g)

gauge_color_g2rd(g)

gauge_color_brbg(g)

gauge_color_piyg(g)

gauge_color_prgn(g)

gauge_color_puor(g)

gauge_color_rdbu(g)

gauge_color_rdgy(g)

gauge_color_rdylbu(g)

gauge_color_rdylgn(g)

gauge_color_spectral(g)
}
\arguments{
\item{g}{An object of class \code{g2r} or \code{g2Proxy} as returned by \code{\link[=g2]{g2()}} or
\code{\link[=g2_proxy]{g2_proxy()}}.}
}
\description{
Convenience function to easily apply colors palettes.
}
\section{Palettes}{


Palletes from the \code{viridisLite} package, ideal for:
continuous data.
\itemize{
\item \code{gauge_color_viridis} (continuous)
\item \code{gauge_color_plasma} (continuous)
\item \code{gauge_color_inferno} (continuous)
\item \code{gauge_color_magma} (continuous)
}

Palettes from color brewer:
\itemize{
\item \code{gauge_color_accent} (qualitative)
\item \code{gauge_color_dark2} (qualitative)
\item \code{gauge_color_paired} (qualitative)
\item \code{gauge_color_pastel1} (qualitative)
\item \code{gauge_color_pastel2} (qualitative)
\item \code{gauge_color_set1} (qualitative)
\item \code{gauge_color_set2} (qualitative)
\item \code{gauge_color_set3} (qualitative)
\item \code{gauge_color_brbg} (diverging)
\item \code{gauge_color_piyg} (diverging)
\item \code{gauge_color_prgn} (diverging)
\item \code{gauge_color_puor} (diverging)
\item \code{gauge_color_rdbu} (diverging)
\item \code{gauge_color_rdgy} (diverging)
\item \code{gauge_color_rdylbu} (diverging)
\item \code{gauge_color_rdylgn} (diverging)
\item \code{gauge_color_spectral} (diverging)
}

Palettes taken from \href{https://coolors.co}{coolors.co}:
\itemize{
\item \code{gauge_color_neon} (continuous)
\item \code{gauge_color_std} (continuous)
\item \code{gauge_color_orange} (continuous)
\item \code{gauge_color_pink} (continuous)
\item \code{gauge_color_lime} (continuous)
\item \code{gauge_color_blue} (continuous)
}

Palettes from \code{awtools} package:
\itemize{
\item \code{gauge_color_aw} (qualitative)
}

Custom:
\itemize{
\item \code{gauge_color_g2rc} (continuous)
\item \code{gauge_color_g2rq} (qualitative)
\item \code{gauge_color_g2rd} (diverging)
}
}

