% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{plan}
\alias{plan}
\title{Mapping}
\usage{
plan(x, y, ...)
}
\arguments{
\item{x, y, ...}{List of name value pairs giving aesthetics to map to
variables. The names for x and y aesthetics are typically omitted because
they are so common; all other aesthetics must be named.}
}
\description{
Mapping aesthetics, similar to \code{ggplot2}'s aesthetics.
}
\section{Aesthetics}{

\itemize{
  \item{\code{x}, \code{y}}
  \item{\code{size}}
  \item{\code{color}}
  \item{\code{shape}}
  \item{\code{opacity}}
  \item{\code{adjust}}
  \item{\code{tooltip}}
  \item{\code{label}}
  \item{\code{style}}
}
}

