% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plane.R
\name{plane}
\alias{plane}
\alias{plane_wrap}
\alias{planes}
\title{Plane}
\usage{
plane_wrap(g2, ..., type = c("list", "rect", "circle", "tree", "mirror",
  "matrix"), sync = TRUE)

planes(...)
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}

\item{...}{Planes, bare column names or \code{NULL}.}

\item{type}{The type of facet to user.}

\item{sync}{Wether to \code{\link{sync}} the \code{planes} and main mapping that were passed \code{\link{g2}}}
}
\description{
This function is similar to the \code{facet_*} family of functions from the \code{ggplot2} package.
}
\examples{
iris \%>\%
  g2(plan(Petal.Length, Petal.Width, color = Species)) \%>\% 
  fig_point() \%>\%
  plane_wrap(planes(Species)) 

}
