% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plane.R
\name{plane}
\alias{plane}
\alias{plane_wrap}
\alias{planes}
\title{Plane}
\usage{
plane_wrap(g2, ..., type = c("list", "rect", "circle", "tree", "mirror",
  "matrix"))

planes(...)
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}

\item{...}{Planes, bare column names or \code{NULL}.}

\item{type}{The type of facet to user.}
}
\description{
This function is similar to the \code{facet_*} family of functions from the \code{ggplot2} package.
}
