% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planes.R
\name{planes}
\alias{planes}
\title{Planes}
\usage{
planes(g, asp, ..., type = c("rect", "list", "matrix", "circle", "tree"))
}
\arguments{
\item{g}{An object of class \code{g2r} or \code{g2Proxy} as returned by \code{\link[=g2]{g2()}} or
\code{\link[=g2_proxy]{g2_proxy()}}.}

\item{asp}{Aspects that define split.}

\item{...}{Any other option.}

\item{type}{Type of planes to use.}
}
\description{
Split the chart into planes according to variables.
}
\examples{
g2(iris, asp(Sepal.Length, Sepal.Width, color = Species)) \%>\% 
 fig_point() \%>\% 
 planes("Species", type = "tree")

}
