% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale-color}
\alias{scale-color}
\alias{g2_scale_colour}
\alias{g2_scale_color}
\title{Scale colour}
\usage{
g2_scale_colour(g2, colors)

g2_scale_color(g2, colors)
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}

\item{colors}{A vector of colors or a JavaScript function.}
}
\description{
Scale colour.
}
