% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slider.R
\name{slider}
\alias{slider}
\title{Slider}
\usage{
slider(g2, ..., position = c("bottom", "top"), id = NULL)
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}

\item{...}{Any option, aesthetic (\code{\link{asp}}), or animation (\code{\link{Animation}}).}

\item{position}{Position of the slider relative to the chart.}

\item{id}{DOM id of slider.}
}
\description{
Create a slider.
}
\examples{
range <- range(cars$speed)

g2(cars, asp(speed, dist)) \%>\% 
  fig_point() \%>\% 
  slider(start = range[1], end = range[2])

}
