% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{sync}
\alias{sync}
\title{Sync}
\usage{
sync(g, position, sync = TRUE)
}
\arguments{
\item{g}{An object of class \code{g2r} or \code{g2Proxy} as returned by \code{\link[=g2]{g2()}} or
\code{\link[=g2_proxy]{g2_proxy()}}.}

\item{position}{Vector of column names of length 2
indicating x and y columns.}

\item{sync}{Either a logical or the name of a group.}
}
\description{
Sets the syncs groups for x and y variables.
}
\keyword{internal}
