% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_options.R
\name{sync}
\alias{sync}
\title{Sync}
\usage{
sync(g2, ...)
}
\description{
Synchronises variables across planes (\code{\link{plane_wrap}}).
}
\examples{
iris \%>\% 
  g2(plan(Sepal.Length, Sepal.Width)) \%>\% 
  fig_point(plan(color = Species)) \%>\% 
  plane_wrap(planes(Species)) \%>\% 
  sync(Sepal.Length, Sepal.Width)

}
