% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_options.R
\name{sync}
\alias{sync}
\title{Sync}
\usage{
sync(g2, ...)
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}

\item{...}{Bare Columng names of variables to synchronise.}
}
\description{
Synchronises variables across planes (\code{\link{plane_wrap}}), in order to have them share scales across
}
\examples{
iris \%>\% 
  g2(asp(Sepal.Length, Sepal.Width)) \%>\% 
  fig_point(asp(color = Species)) \%>\% 
  plane_wrap(planes(Species)) \%>\% 
  sync(Sepal.Length, Sepal.Width)

}
