% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\name{theme}
\alias{theme}
\alias{motif_dark}
\alias{custom_motif}
\title{Motif}
\usage{
motif_dark(g2)

custom_motif(g2, ...)
}
\arguments{
\item{g2}{An object of class \code{g2r} as returned by \code{\link{g2r}}.}

\item{...}{Motif options.}
}
\description{
Customise the chart motif.
}
\examples{
g <- g2(iris, asp(Sepal.Length, Sepal.Width)) \%>\% 
  fig_point(asp(color = Species)) 

g \%>\% motif_dark()

g \%>\% custom_motif(colors = c("red", "green", "blue"))

}
